% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignment.signal.R
\name{alignment.signal}
\alias{alignment.signal}
\title{Rboretum Alignment Signal Fetcher}
\usage{
alignment.signal(
  alignment_path,
  species_info,
  informative_gaps,
  alignment_name
)
}
\arguments{
\item{alignment_path}{Path to alignment file (absolute or relative)}

\item{species_info}{Can be EITHER (1) A phylo object from which species will be extracted; or (2) a character vector of desired taxa (> 3 species required)}

\item{informative_gaps}{If TRUE, gaps in the alignment (-) are treated as potentially informative indels. If FALSE, gaps are considered missing data and not used (Default: FALSE)}

\item{alignment_name}{Name for alignment. If missing, the base filename is used}
}
\value{
Dataframe containing split pattern for each site in the alignment, relative to the given set of taxa
}
\description{
Given the path to an alignment and list of taxa, this script returns site patterns for each site in the alignment (after pruning if needed)
}
\examples{
myAlignPath <- '/path/to/align.phy'

mySpecies <- treeToCheck # Phylo object with species of interest
OR
mySpecies <- c('Spp1','Spp2','Spp3'...)

# Gaps are indels
alignment.signal(myAlignPath,mySpecies,informative_gaps = TRUE)

# Gaps are missing data
alignment.signal(myAlignPath,mySpecies,informative_gaps = FALSE)

# Add alignment name
alignment.signal(myAlignPath,mySpecies,informative_gaps = FALSE, alignment_name = 'gene_XYZ')

}
