% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicTreePlot.R
\name{basicTreePlot}
\alias{basicTreePlot}
\title{Basic Tree Plotter}
\usage{
basicTreePlot(tree, label, xmax)
}
\arguments{
\item{label}{Either 'node' [print node labels] or 'bs' [print bootstrap labels]. Any other entry (or missing) results in no node labels}

\item{xmax}{Optional: Provide a numeric value to extend plot axis (i.e. if right side is cut off)}

\item{tree_1}{Phylo object}
}
\description{
This function plots a tree without branch lengths, and nodes can be labeled with bootstrap or node labels
}
\examples{
# Print tree with no labels
basicTreePlot(tree)

# Print tree with bootstrap labels
basicTreePlot(tree,'bs')

# Print tree with node labels
basicTreePlot(tree,'node')

}
