% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basicTreePlotter2.R
\name{basicTreePlotter2}
\alias{basicTreePlotter2}
\title{Rboretum Basic Tree Plotter}
\usage{
basicTreePlotter2(
  tree,
  branch_length,
  branch_weight,
  node_label,
  node_size,
  node_nudge,
  taxa_size,
  taxa_italic,
  taxa_align,
  taxa_offset,
  xmax,
  reverse_x,
  to_color,
  colors,
  color_legend
)
}
\arguments{
\item{tree}{Tree(s) to plot. Options include:
\itemize{
  \item A single, rooted phylo object [Clade support disabled]; or, 
  \item A named, rooted multiPhylo object [Only unique topologies plotted]
}}

\item{branch_length}{OPTIONAL: TRUE [plot tree(s) with branch lengths]; FALSE [Default: plot cladogram(s)]}

\item{branch_weight}{OPTIONAL: Set ggtree branch thickness}

\item{node_label}{OPTIONAL: Node label choices include:
\itemize{
  \item "none": No node labels
  \item "node": Node number
  \item "bs": Bootstrap value [Default]
}}

\item{node_size}{OPTIONAL: Set ggtree node label size}

\item{node_nudge}{OPTIONAL: Set ggtree node label nudge_x}

\item{taxa_size}{OPTIONAL: Set ggtree tip label size}

\item{taxa_italic}{OPTIONAL: TRUE [italic tip labels]; FALSE [Default: non-italic tip labels]}

\item{taxa_align}{OPTIONAL: 'left' or 'right' tip label alignment [Default: No alignment]}

\item{taxa_offset}{OPTIONAL: Set ggtree tip label offset}

\item{xmax}{OPTIONAL: Set ggplot xlim upper limit (e.g if long tip labels run off plot)}

\item{reverse_x}{OPTIONAL: TRUE [plot tree with tips on left]; FALSE [Default: plot tree with tips on right]}

\item{to_color}{OPTIONAL: Color tips or clades via:
\itemize{
  \item Character vector of taxa, all to be colored the same color
  \item List of groups of taxa, each of which will have their own color. List can be named for use with a legend (set color_legend == TRUE)
}}

\item{colors}{OPTIONAL: Colors for clade highlighting. Must be hex or valid R colors. Provide a color for each group (1 if character vector, 1 for each group if named list) or default colors will be used.}

\item{color_legend}{OPTIONAL: TRUE [Include a legend for colored taxa/clades]; False [Default: No legend]}
}
\value{
ggtree object
}
\description{
Given a phylo or multiPhylo object (tree), this ggtree wrapper returns a ggtree plot object, adjusted with possible arguments
}
