% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_getAlignmentSignal.R
\name{batch_getAlignmentSignal}
\alias{batch_getAlignmentSignal}
\title{Rboretum Batch Alignment Signal Fetcher}
\usage{
batch_getAlignmentSignal(
  alignment_paths,
  species_info,
  use_gaps,
  alignment_names
)
}
\arguments{
\item{alignment_paths}{Vector of paths to alignment files (absolute or relative)}

\item{species_info}{Set of 3+ species that can be passed as either:
\itemize{
  \item phylo object from which species will be extracted; or
  \item Character vector of desired taxa
}}

\item{use_gaps}{\itemize{
  \item FALSE (Default: Treat all gaps (-) in all alignments as missing data)
  \item TRUE (Treat all gaps in all alignments as potentially informative signal)
  \item Logical vecor indicating TRUE or FALSE for each alignment
}}

\item{alignment_names}{Chacter vector of names for each alignment. If missing or incomplete, the base filename is used}
}
\value{
Dataframe containing split pattern for each site, in each alignment, relative to the given set of taxa
}
\description{
Given the paths to multiple alignments and a common list of taxa, this script returns site patterns for each site, in each alignment (after pruning if needed)
}
