% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_readRootedTree.R
\name{batch_readRootedTree}
\alias{batch_readRootedTree}
\title{Rboretum Multiphylo Rooted Tree Reader}
\usage{
batch_readRootedTree(tree_paths, root_taxa)
}
\arguments{
\item{tree_paths}{Vector of paths to tree files that can be read by ape::read.tree() or ape::read.nexus()}

\item{root_taxa}{Character vector containing outgroup species IDs (Must be in all trees and always monophyletic)}
}
\value{
A multiPhylo object, with each tree rooted at specified taxa
}
\description{
This function is an ape wrapper, and returns a multiPhylo object, with each tree rooted as specified by the user.
}
\examples{
myTrees <- c('/path/to/tree1','/path/to/tree2')
myRootTaxa <- c('Spp1','Spp2')
myMulti <- readMulti.rooted(myTrees,myRootTaxa)
}
