% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canCompare.R
\name{canCompare}
\alias{canCompare}
\title{Rboretum Multiphylo Comparability Checker}
\usage{
canCompare(trees, tree_names, return_only_comparable)
}
\arguments{
\item{trees}{Multiphylo object}

\item{tree_names}{OPTIONAL: Vector of tree names}

\item{return_only_comparable}{OPTIONAL: If TRUE, returns only comparable trees; (Default: FALSE)}
}
\value{
Dataframe with information about possible comparisons
}
\description{
This function takes a multiphylo object and returns a dataframe that indicates which topologies can be compared (>= 3 species shared, with unique topologies). If so, the common species set is also returned.
}
\examples{

trees <- c(tree_1,tree_2,tree_3)
names <- c('CDS_Tree','Intron_Tree','UTR_Tree')

# Return all comparisons
canCompare(trees,names)

# Return only comparable trees
canCompare(trees,names,return_only_comparable=TRUE)

}
