% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_name.default.R
\name{change_name.default}
\alias{change_name.default}
\title{Add New Scale
Source: eliocamp/new_aes.R}
\usage{
\method{change_name}{default}(list, old, new)
}
\description{
Add New Scale
Source: eliocamp/new_aes.R
}
