% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_name.NULL.R
\name{change_name.NULL}
\alias{change_name.NULL}
\title{Add New Scale
Source: eliocamp/new_aes.R}
\usage{
\method{change_name}{`NULL`}(list, old, new)
}
\description{
Add New Scale
Source: eliocamp/new_aes.R
}
