% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.root.R
\name{check.root}
\alias{check.root}
\title{Rboretum Root Checker}
\usage{
check.root(tree, taxa)
}
\arguments{
\item{tree}{Rooted phylo object}

\item{taxa}{Character vector containing taxa to check}
}
\value{
TRUE if 'taxa' are either side of the root, else, FALSE
}
\description{
This function returns TRUE if  'taxa' are part of a root split
}
\examples{
check.root(tree,taxa_to_check)

}
