% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkComparable.R
\name{checkComparable}
\alias{checkComparable}
\title{Rboretum Two-Tree Comparability Checker}
\usage{
checkComparable(tree_1, tree_2)
}
\arguments{
\item{tree_1}{phylo object}

\item{tree_2}{phylo object}
}
\value{
TRUE if 'tree_1' and 'tree_2' have 3 or more species in common and a unique topology; else, FALSE
}
\description{
This function returns TRUE if 'tree_1' and 'tree_2' have 3 or more species in common, and a unique topology
}
