% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkComparableMulti.R
\name{checkComparableMulti}
\alias{checkComparableMulti}
\title{Check Which Tree Topologies from Multiphylo Can Be Compared}
\usage{
checkComparableMulti(trees, names, return_only_comparable)
}
\arguments{
\item{trees}{Multiphylo object}

\item{names}{Vector of tree identifiers (must have identifier for each tree)}

\item{return_only_comparable}{OPTIONAL: If TRUE, returns only comparable trees; Anything else, including missing, returns all comparisons (Default: FALSE)}
}
\value{
Dataframe with information about possible comparisons
}
\description{
This function takes a multiphylo object and returns a dataframe that indicates which topologies can be compared (>= 3 species shared, with unique topologies). If so, the common species set is also returned.
}
\examples{

trees <- c(tree_1,tree_2,tree_3)
names <- c('CDS_Tree','Intron_Tree','UTR_Tree')

# Return all comparisons
checkComparableMulti(trees,names)

# Return only comparable trees
checkComparableMulti(trees,names,return_only_comparable=TRUE)

}
