% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIdenticalSpecies.R
\name{checkIdenticalSpecies}
\alias{checkIdenticalSpecies}
\title{Check if Trees Share All Species}
\usage{
checkIdenticalSpecies(tree_1, tree_2)
}
\arguments{
\item{tree_1}{Phylo object}

\item{tree_2}{Phylo object}
}
\value{
TRUE if 'tree_1' and 'tree_2' have all species in commmon; else, FALSE
}
\description{
This function returns TRUE if 'tree_1' and 'tree_2' have all species in commmon
}
\examples{
checkIdenticalSpecies(tree_1,tree_2)

}
