% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIdenticalSpeciesMulti.R
\name{checkIdenticalSpeciesMulti}
\alias{checkIdenticalSpeciesMulti}
\title{Check for Identical Species Among Trees in Multiphylo}
\usage{
checkIdenticalSpeciesMulti(trees)
}
\arguments{
\item{trees}{Multiphylo object}
}
\value{
TRUE if all trees have all species in common; else, FALSE
}
\description{
This function takes a multiphylo object and returns TRUE if all species are shared among all trees; otherwise, FALSE
}
\examples{
trees <- c(tree_1,tree_2,tree_3,...)
checkIdenticalSpeciesMulti(trees)

}
