% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMonophyletic.R
\name{checkMonophyletic}
\alias{checkMonophyletic}
\title{Check Monophyly of Taxa}
\usage{
checkMonophyletic(tree, taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{taxa}{Vector containing taxa to check}
}
\value{
TRUE if 'taxa' monophyletic in 'tree', else, FALSE
}
\description{
This function returns TRUE if  'taxa' are monophyletic in 'tree', FALSE otherwise
}
\examples{
checkMonophyletic(tree,taxa_to_check)

}
