% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSameTaxa.R
\name{checkSameTaxa}
\alias{checkSameTaxa}
\title{Rboretum Identical Taxon Checker}
\usage{
checkSameTaxa(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
TRUE if all trees have all species in common; else, FALSE
}
\description{
This function takes a multiPhylo object and returns TRUE if all species are shared among all trees; otherwise, FALSE
}
