% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSameTopology.R
\name{checkSameTopology}
\alias{checkSameTopology}
\title{Rboretum Identical Topology Checker}
\usage{
checkSameTopology(trees, check_any)
}
\arguments{
\item{trees}{Rooted, multiPhylo object where all trees share at least three taxa}

\item{check_any}{OPTIONAL: If TRUE, check if ANY (rather than all) trees share a topology [Default: FALSE, check for unanimous topology]}
}
\value{
TRUE if all trees have the same topology after pruning to identical species lists; else, FALSE
}
\description{
This function returns TRUE if all trees in multiPhylo have the same topology after pruning to identical species lists
}
\examples{
checkSameTopology(myTrees) # Check if all trees in multiPhylo 'myTrees' share the same topology
checkSameTopology(myTrees,check_any=TRUE) # Check if any trees in multiPhylo 'myTrees' share a topology
}
