% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSharedSpecies.R
\name{checkSharedSpecies}
\alias{checkSharedSpecies}
\title{Check for Shared Species}
\usage{
checkSharedSpecies(tree_1, tree_2)
}
\arguments{
\item{tree_1}{Phylo object}

\item{tree_2}{Phylo object}
}
\value{
TRUE if 'tree_1' and 'tree_2' have 3 or more  species in commmon; else, FALSE
}
\description{
This function returns TRUE if 'tree_1' and 'tree_2' have 3 or more species in commmon
}
\examples{
checkSharedSpecies(tree_1,tree_2)

}
