% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSharedSpeciesMulti.R
\name{checkSharedSpeciesMulti}
\alias{checkSharedSpeciesMulti}
\title{Check for 3+ Shared Species Among Trees in Multiphylo}
\usage{
checkSharedSpeciesMulti(trees)
}
\arguments{
\item{trees}{Multiphylo object}
}
\value{
TRUE (all trees share >= 3 species) or FALSE
}
\description{
This function takes a multiphylo object and returns TRUE if 3 or more species are shared among all trees; otherwise, FALSE
}
\examples{
trees <- c(tree_1,tree_2,tree_3)
checkSharedSpeciesMulti(trees)

}
