% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSharedTaxa.R
\name{checkSharedTaxa}
\alias{checkSharedTaxa}
\title{Rboretum Three Common Taxa Checker}
\usage{
checkSharedTaxa(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
TRUE (all trees share >= 3 species) or FALSE
}
\description{
This function takes a multiPhylo object and returns TRUE if all trees share at least three common taxa (e.g. enough to make a tree); otherwise, FALSE
}
