% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTreeTaxa.R
\name{checkTreeTaxa}
\alias{checkTreeTaxa}
\title{Check Tree Taxa}
\usage{
checkTreeTaxa(tree, taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{taxa}{Vector containing taxa to check}
}
\value{
TRUE if 'taxa' in 'tree', else, FALSE
}
\description{
This function returns TRUE if all 'taxa' are present in 'tree', FALSE otherwise
}
\examples{
checkTreeTaxa(tree,taxa_to_find)

}
