% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTreeTaxaMulti.R
\name{checkTreeTaxaMulti}
\alias{checkTreeTaxaMulti}
\title{Rboreturm Multiphlyo Taxon Checker}
\usage{
checkTreeTaxaMulti(trees, taxa)
}
\arguments{
\item{trees}{Multiphylo object}

\item{taxa}{Character vector containing taxa to query}
}
\value{
TRUE if all 'taxa' in all 'trees', else, FALSE
}
\description{
This function returns TRUE if all 'taxa' are present in all 'trees', FALSE otherwise
}
\examples{

myTrees <- c(treeA,treeB,treeC)
taxa_to_find <- c('Spp1','Spp2','Spp3')
checkTreeTaxa(myTrees,taxa_to_find)

}
