% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade.support.R
\name{clade.support}
\alias{clade.support}
\title{Rboretum Clade Support Fetcher}
\usage{
clade.support(
  signal,
  clade,
  max_missing,
  include_gap,
  include_singleton,
  include_biallelic,
  include_triallelic,
  include_quadallelic,
  include_pentallelic,
  only_gap,
  as_root
)
}
\arguments{
\item{signal}{Output table from alignment.signal()}

\item{clade}{Character vector containing all taxa in clade of interest [Note: All requested taxa must appear in alignment signal data]}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column [Default: 0]}

\item{include_gap}{OPTIONAL: TRUE or FALSE; Count sites with gap positions ('-') as part of total support [Default: TRUE]}

\item{include_singleton}{OPTIONAL: TRUE or FALSE; Count sites with singletons as part of total support [Default: TRUE]}

\item{include_biallelic}{OPTIONAL: TRUE or FALSE; Count sites with biiallelic variation as part of total support [Default: TRUE]}

\item{include_triallelic}{OPTIONAL: TRUE or FALSE; Count sites with triallelic variation as part of total support [Default: TRUE]}

\item{include_quadallelic}{OPTIONAL: TRUE or FALSE; Count sites with quadallelic variation as part of total support [Default: TRUE]}

\item{include_pentallelic}{OPTIONAL: TRUE or FALSE; Count sites with pentallelic variation as part of total support [Default: TRUE]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') as part of total support [Default: FALSE]}

\item{as_root}{OPTIONAL: Process signal as if it were a root split (biallelic only + no missing taxa allowed) [Default: FALSE]}
}
\value{
integer count of total sites of support for clade in available signal
}
\description{
This function takes a character vector argument of taxon, and returns the total site support count for that clade from an alignment [output from alignment.signal()]
}
\examples{
clade.support <- (signal,clade)

}
