% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.clades.R
\name{compare.clades}
\alias{compare.clades}
\title{Rboretum Tree Clade Comparison}
\usage{
compare.clades(trees, tree_names, return_shared_only)
}
\arguments{
\item{trees}{multiPhylo object}

\item{tree_names}{OPTIONAL: Vector of tree names}

\item{return_shared_only}{OPTIONAL: If TRUE, returns only information about clades supported by all trees; (Default: FALSE)}
}
\value{
Dataframe with clade support information
}
\description{
This function takes a multiPhylo object, and queries support for all uniquely identified monophyletic clades
}
\examples{
trees <- c(tree_1,tree_2,tree_3)

# Return information on all possible splits
compare.clades(trees)

# Return information on only those splits shared in all trees
compare.clades(trees,return_shared_only=TRUE)

}
