% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareClades.R
\name{compareClades}
\alias{compareClades}
\title{Rboretum Tree Clade Comparison}
\usage{
compareClades(trees)
}
\arguments{
\item{trees}{Named multiPhylo object [Set tree names by using names(trees) <- c('Tree1_Name','Tree2_Name',etc.)]}
}
\value{
Dataframe with clade support information
}
\description{
This function takes a named multiPhylo object, and queries support for all uniquely identified monophyletic clades
}
