% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareShuffledMulti.R
\name{compareShuffledMulti}
\alias{compareShuffledMulti}
\title{Check Split Occurences Among Trees Containing Identical Species}
\usage{
compareShuffledMulti(trees, tree_names, return_shared_only)
}
\arguments{
\item{trees}{Multiphylo object}

\item{tree_names}{Vector of tree names}

\item{return_shared_only}{OPTIONAL: If TRUE, returns only splits shared in all trees; Anything else, including missing, returns all splits and their counts (Default: FALSE)}
}
\value{
Dataframe with information about splits among trees
}
\description{
This function takes a multiphylo object where each tree contains identical taxa, and returns a summary of split occurrences
}
\examples{
trees <- c(tree_1,tree_2,tree_3)

# Return information on all possible splits
compareShuffledMulti(trees)

# Return information on only those splits shared in all trees
compareShuffledMulti(trees,return_shared_only=TRUE)

}
