% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareSplits.R
\name{compareSplits}
\alias{compareSplits}
\title{Compare Splits for Two Unique Topologies}
\usage{
compareSplits(tree_1, tree_2)
}
\arguments{
\item{tree_1}{Phylo object}

\item{tree_2}{Phylo object}
}
\value{
Dataframe with all splits in tree_1 and tree_2, with common and unique splits noted in the 'Split_Group' column [Shared,Unique_Tree_1,Unique_Tree_2]
}
\description{
This function returns a dataframe with all splits in tree_1 and tree_2, with common and unique splits noted in the 'Split_Group' column [Shared,Unique_Tree_1,Unique_Tree_2]
}
\examples{
compareSplits(tree_1,tree_2)

}
