% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getComparable.R
\name{compareTrees}
\alias{compareTrees}
\title{Rboretum Tree Comparability Checker}
\usage{
compareTrees(trees)
}
\arguments{
\item{trees}{Named multiPhylo object [Set tree names by using names(trees) <- c('Tree1_Name','Tree2_Name',etc.)]}
}
\value{
Dataframe with information about possible comparisons
}
\description{
This function takes a multiPhylo object and returns a dataframe that indicates which topologies can be compared (>= 3 species shared, with unique topologies). If so, the common species set is also returned.
}
