% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.counter.R
\name{concordance.counter}
\alias{concordance.counter}
\title{Rboretum Concordance Counter}
\usage{
concordance.counter(signal, tree)
}
\arguments{
\item{signal}{Output table from getAlignmentSignal()}

\item{tree}{Rooted phylo object}
}
\value{
Dataframe containing the subset of 'signal' that has split information, and the count of concordant splits for each site
}
\description{
This function takes the signal from an alignment file [via getAlignmentSignal()] and counts the number of splits that provide concordant signal, relative to a given tree
}
