% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.tips.R
\name{convert.tips}
\alias{convert.tips}
\title{Rboretum Tip Label Swapper}
\usage{
convert.tips(tree, name_df, from, to)
}
\arguments{
\item{tree}{phylo object}

\item{name_df}{Dataframe with name equivalencies}

\item{from}{Column name with current tree tip IDs}

\item{to}{Column name with desired IDs}
}
\value{
phylo object with converted names
}
\description{
This function takes a tree and changes tip labels based on a supplied dataframe
}
\examples{
convert.tips(tree,name_df,from,to)

}
