% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLabels.R
\name{convertLabels}
\alias{convertLabels}
\title{Rboretum Clade Label Converter}
\usage{
convertLabels(to_convert, name_df, from, to)
}
\arguments{
\item{to_convert}{R object containing taxon labels to be replaced. Options include: 
\itemize{
  \item A phylo object; or,
  \item A multiPhylo object; or,
  \item A character vector of taxon IDs [e.g. c('Species1','Species2',etc.)]; or,
  \item A list of character vectors containing taxon IDs [e.g. list(A=c('Species1','Species2'),B=c('Species_3','Species_4'))]; or,
  \item A character vector of semicolon-separated clades [e.g. c('Species_1;Species_2','Species_3;Species_4')]
}}

\item{name_df}{Dataframe with name equivalencies}

\item{from}{Column name with all current taxon tip IDs (if missing, first column of name_df used as default)}

\item{to}{Column name with desired IDs (if missing, second column of name_df used as default)}
}
\value{
Matching object with new names in place
}
\description{
This function takes a vector of semicolon-separated clades and changes names based on a supplied dataframe
}
