% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTipLabels.R
\name{convertTipLabels}
\alias{convertTipLabels}
\title{Convert Tip Labels}
\usage{
convertTipLabels(tree, name_df, from, to)
}
\arguments{
\item{tree}{Phylo object}

\item{name_df}{Dataframe with name equivalencies}

\item{from}{Column name with current tree tip IDs}

\item{to}{Column name with desired IDs}
}
\value{
Tree objecct with converted names
}
\description{
This function takes a tree and changes tip labels based on a supplied dataframe
}
\examples{
convertTipLabels(tree,name_df,from,to)

}
