% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectSingleRoot.R
\name{detectSingleRoot}
\alias{detectSingleRoot}
\title{Rboreturm Single Root Checker}
\usage{
detectSingleRoot(tree, return_root)
}
\arguments{
\item{tree}{Tree to check. Must be a rooted phylo object}

\item{return_root}{OPTIONAL: If TRUE and tree is rooted on a single taxon, return that tip label; else return FALSE [Default: FALSE, don't return root label]}
}
\value{
TRUE if 'tree' is rooted on a single taxon; else, FALSE
}
\description{
This function returns TRUE if 'tree' is rooted on a single taxon, or FALSE if not
}
