% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNodeAges.R
\name{extractNodeAges}
\alias{extractNodeAges}
\title{Rboretum Node Age Extractor}
\usage{
extractNodeAges(tree, return_summary)
}
\arguments{
\item{tree}{Tree(s) to extract date information from. Options include:
\itemize{
  \item A single, rooted phylo object; or,
  \item A rooted multiPhylo object where all trees share 3+ taxa and support a single topology 
}}

\item{return_summary}{OPTIONAL: If TRUE and a multiPhylo is provided, return a summary of node ages across trees with mean and median values per clade}
}
\value{
A dataframe containing (a) node/date information for each tree or (b) clade-level ages summarized across trees in a multiPhlyo if return_summary=TRUE
}
\description{
This function takes a tree(s) where branches are scaled to time and returns the estimated node ages, or a summary of node ages across trees
}
