% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.comparable.R
\name{get.comparable}
\alias{get.comparable}
\title{Rboretum Tree Comparability Checker}
\usage{
get.comparable(trees, return_only_comparable)
}
\arguments{
\item{trees}{multiPhylo object (If named, names will be used in analysis)}

\item{return_only_comparable}{OPTIONAL: If TRUE, returns only comparable trees; (Default: FALSE)}
}
\value{
Dataframe with information about possible comparisons
}
\description{
This function takes a multiPhylo object and returns a dataframe that indicates which topologies can be compared (>= 3 species shared, with unique topologies). If so, the common species set is also returned.
}
\examples{

trees <- c(tree_1,tree_2,tree_3)
names <- c('CDS_Tree','Intron_Tree','UTR_Tree')

# Return all comparisons
get.comparable(trees,names)

# Return only comparable trees
get.comparable(trees,names,return_only_comparable=TRUE)

}
