% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.shared.R
\name{get.shared}
\alias{get.shared}
\title{Rboretum Common Taxa Fetcher}
\usage{
get.shared(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
Sorted character vector of tip labels present in all trees
}
\description{
This function takes a multiPhylo object and returns a sorted list of taxa common to all trees. Note: Function will STOP if no taxa are shared.
}
\examples{
trees <- c(tree_1,tree_2,tree_3)
get.shared(trees)

}
