% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.splits.R
\name{get.splits}
\alias{get.splits}
\title{Rboretum Tree Splitter}
\usage{
get.splits(tree)
}
\arguments{
\item{tree}{Rooted phylo object}
}
\value{
Four-column dataframe; 1: Semi-colon separated monophyletic clade; 2: 'Mirror Clade'; 3: Phylo Node ID; 4: Node Boostrap (if present)
}
\description{
This function takes a rooted phylo object and returns information about each monophyletic group/split
}
\examples{
get.splits(tree)

}
