% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.uniqueClades.R
\name{get.uniqueClades}
\alias{get.uniqueClades}
\title{Rboretum Unique Clade Fetcher}
\usage{
get.uniqueClades(trees, focal_tree)
}
\arguments{
\item{trees}{Named multiPhylo object}

\item{focal_tree}{Name of focal tree}
}
\value{
Character vector of clades that occur in the focal tree, and are missing from at least one other tree in the multiPhylo
}
\description{
This function takes a named multiPhylo object and a focal tree, and returns clades that occur in the focal clade but not all other trees in the multiPhylo
}
