% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentSignal.R
\name{getAlignmentSignal}
\alias{getAlignmentSignal}
\title{Rboretum Alignment Signal Fetcher}
\usage{
getAlignmentSignal(alignment_path, species_info, use_gaps, alignment_name)
}
\arguments{
\item{alignment_path}{Path to alignment file (absolute or relative)}

\item{species_info}{Set of 3+ species that can be passed as either:
\itemize{
  \item phylo object from which species will be extracted; or
  \item Character vector of desired taxa
}}

\item{use_gaps}{OPTIONAL: Options include:
\itemize{
  \item FALSE (Default: Treat all gaps (-) in alignment as missing data)
  \item TRUE (Treat all gaps in alignment as potentially informative signal)
}}

\item{alignment_name}{OPTIONAL: Name for alignment. If missing, the base filename is used}
}
\value{
Dataframe containing split pattern for each site in the alignment, relative to the given set of taxa
}
\description{
Given the path to an alignment and list of taxa, this script returns site patterns for each site in the alignment (after pruning if needed)
}
