% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentSignal_Worker.R
\name{getAlignmentSignal_Worker}
\alias{getAlignmentSignal_Worker}
\title{Rboretum Alignment Signal Worker}
\usage{
getAlignmentSignal_Worker(
  alignment_path,
  species_info,
  use_gaps,
  alignment_name
)
}
\arguments{
\item{alignment_path}{Path to alignment file (absolute or relative)}

\item{species_info}{List of taxa to analyze. Can be provided as:
\itemize{
  \item phylo object from which species will be extracted; or
  \item multiPhylo object from which common species will be extracted; or
  \item Character vector with 3+ taxon IDs
}}

\item{use_gaps}{Options include:
\itemize{
  \item FALSE (Default: Treat all gaps (-) in alignment as missing data)
  \item TRUE (Treat all gaps in alignment as potentially informative signal)
}}

\item{alignment_name}{Name for alignment. [Default: File base name]}
}
\value{
Dataframe containing split pattern for each site in the alignment, relative to the given set of taxa
}
\description{
Given the path to an alignment and list of taxa, this script returns site patterns for each site in the alignment (after pruning if needed)
}
