% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentSpecies.R
\name{getAlignmentSpecies}
\alias{getAlignmentSpecies}
\title{Rboretum Alignment Species Fetcher}
\usage{
getAlignmentSpecies(alignment_path)
}
\arguments{
\item{alignment_path}{Relative or absolute path to alignment file}
}
\value{
Sorted, semicolon-separated list of taxa
}
\description{
Given the path to an alignment, this function returns the taxon labels as a sorted, semicolon-separated character
}
