% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSplits.R
\name{getAllSplits}
\alias{getAllSplits}
\title{Get All Splits}
\usage{
getAllSplits(rooted_tree, exclude_root)
}
\arguments{
\item{rooted_tree}{Rooted phylo object}

\item{exclude_root}{OPTIONAL: If TRUE, exclude root clades. Otherwise, return all groups (Default: FALSE)}
}
\value{
Three-column dataframe; 1: Monophyletic clade; 2: 'Mirror Clade'; 3: Node ID
}
\description{
This function takes a rooted tree and returns information about each monophyletic group/split
}
\examples{
getAllSplits(rooted_tree)

}
