% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getChronosDates.R
\name{getChronosDates}
\alias{getChronosDates}
\title{Get Chronos Node Dates}
\usage{
getChronosDates(tree, iter_count, min_root, max_root)
}
\arguments{
\item{tree}{Phylo object}

\item{iter_count}{Number of iterations for chronos}

\item{min_root}{Minimum divergence time estimate for root node}

\item{max_root}{Maximum divergence time estimate for root node}
}
\value{
Ultrametric chronogram
}
\description{
This function takes a tree and root node age calibrations, and converts branch lengths into divergence times using Chronos in ape
}
\examples{

# If root node is forced at 50MYA, with 2000 iterations
getChronosDates(tree,2000,50,50)

# If root node is estimated between 30 and 50 MYA, with 1000 iterations
getChronosDates(tree,1000,30,50)

}
