% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRoot.R
\name{getRoot}
\alias{getRoot}
\title{Rboretum Root Fetcher}
\usage{
getRoot(tree)
}
\arguments{
\item{tree}{A rooted phylo object}
}
\value{
A two-element character vector containing semicolon-delimited root clades
}
\description{
This function takes a rooted phylo object, and returns the representative root split as two semicolon-delimited characters
}
