% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharedSpecies.R
\name{getSharedSpecies}
\alias{getSharedSpecies}
\title{Get Shared Species From Two Trees}
\usage{
getSharedSpecies(tree_1, tree_2)
}
\arguments{
\item{tree_1}{Phylo object}

\item{tree_2}{Phylo object}
}
\value{
Vector of species shared between 'tree_1' and 'tree_2
}
\description{
Returns sorted list of species that occur in 'tree_1' and 'tree_2'
}
\examples{
getSharedSpecies(tree_1,tree_2)

}
