% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharedSpeciesMulti.R
\name{getSharedSpeciesMulti}
\alias{getSharedSpeciesMulti}
\title{Get Shared Species List from Multiphylo Object}
\usage{
getSharedSpeciesMulti(trees)
}
\arguments{
\item{trees}{Multiphylo object}
}
\value{
Sorted vector of taxa shared among all trees in multiphylo
}
\description{
This function takes a multiphylo object and returns a sorted vector of taxa shared among all trees
}
\examples{
trees <- c(tree_1,tree_2,tree_3)

getSharedSpeciesMulti(trees)

}
