% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharedTaxa.R
\name{getSharedTaxa}
\alias{getSharedTaxa}
\title{Rboretum Common Taxa Fetcher}
\usage{
getSharedTaxa(trees)
}
\arguments{
\item{trees}{multiPhylo object where >= 3 species are shared among all trees}
}
\value{
Sorted character vector of tip labels present in all trees
}
\description{
This function takes a multiPhylo object and returns a sorted list of taxa common to all trees. Note: Function will STOP if <3 taxa are shared among trees.
}
\examples{
common_taxa <- getSharedTaxa(myMultiPhylo)
}
