% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getShuffledSupport.R
\name{getShuffledSupport}
\alias{getShuffledSupport}
\title{Plot Alignment Support for Specific Topology}
\usage{
getShuffledSupport(comparison, signal, existing_support)
}
\arguments{
\item{signal}{Output table from getAlignmentSignal run with same taxa as in tree}

\item{existing_support}{OPTIONAL: Should be output of getShuffledSupport from the same comparison, run with different signal}
}
\value{
Tree with nodepoint geoms scaled to support, and colored by number of trees containing that split
}
\description{
DESCRIPTION
}
\examples{
getShuffledSupport(comparison,signal,existing_support)

}
