% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSignalCounts.R
\name{getSignalCounts}
\alias{getSignalCounts}
\title{Rboretum Alignment Signal Support Tabler}
\usage{
getSignalCounts(
  signal,
  clade,
  separate,
  max_missing,
  include_gap,
  only_gap,
  include_singleton,
  include_biallelic,
  include_triallelic,
  include_quadallelic,
  include_pentallelic
)
}
\arguments{
\item{signal}{Output table from getAlignmentSignal()}

\item{clade}{OPTIONAL: Character vector containing all members (2+) of clade for which support is requested. If provided, return total intger support for this set of taxa, rather than the entire table.}

\item{separate}{OPTIONAL: If true, separate results from different alignments (as coded by the Alignment_Name column) [Default: FALSE, return a single table/count]}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column before it is not considered [Default: 0, no missing taxa allowed]}

\item{include_gap}{OPTIONAL: If TRUE, count sites with gap positions ('-') as informative signal; otherwise, count gaps as missing data [Default: FALSE: Gaps are treated as missing]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') [Default: FALSE]}

\item{include_singleton}{OPTIONAL: If TRUE, count sites with singletons as part of total support [Default: TRUE]}

\item{include_biallelic}{OPTIONAL: If TRUE, count sites with biiallelic variation as part of total support [Default: TRUE]}

\item{include_triallelic}{OPTIONAL: If TRUE, count sites with triallelic variation as part of total support [Default: TRUE]}

\item{include_quadallelic}{OPTIONAL: If TRUE, count sites with quadallelic variation as part of total support [Default: TRUE]}

\item{include_pentallelic}{OPTIONAL: If TRUE, count sites with pentallelic variation as part of total support [Default: TRUE]}
}
\value{
Table of counts for clade support among  all sites in multiple-sequence alignment; or support for a specific clade if requested.
}
\description{
This function extracts phylogenetic signal from a multiple-sequence alignment and returns a table where the names are semicolon-separated supported clades
}
