% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSpeciesComposition.R
\name{getSpeciesComposition}
\alias{getSpeciesComposition}
\title{Rboretum Species Composition Fetcher}
\usage{
getSpeciesComposition(alignment_path, alignment_name, prefix, suffix)
}
\arguments{
\item{alignment_path}{Where to find alignment files. Options include:
\itemize{
  \item A character vector of one or more alignment file paths  (relative or absolute)
  \item A path to a single directory containing all alignment files (relative or absolute)
}}

\item{alignment_name}{OPTIONAL: Chacter vector of names for each alignment. If missing or incomplete, the base filename is used}

\item{prefix}{OPTIONAL: If 'alignment_path' is a directory, provide a character vector of file prefixes (e.g. all alignment files start with "Mafft_")}

\item{suffix}{OPTIONAL: If 'alignment_path' is a directory, provide a character vector of file suffixes (e.g. all alignment files end with ".phy")}
}
\value{
Dataframe containing %GC, %N, and %Gap for each taxon in the alignment(s)
}
\description{
Given the path(s) to multiple alignments, this script returns  %GC, %N, and %Gap for each taxon
}
