% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplitBootstrap.R
\name{getSplitBootstrap}
\alias{getSplitBootstrap}
\title{Get Split Bootstrap}
\usage{
getSplitBootstrap(tree, taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{taxa}{Vector containing taxa to check}
}
\value{
Bootstrap support label for MRCA of 'taxa' in 'tree'
}
\description{
This function returns the boostrap support for the MRCA of 'taxa' in 'tree'
}
\examples{
getSplitBootstrap(tree,taxa)

}
