% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplitNode.R
\name{getSplitNode}
\alias{getSplitNode}
\title{Get Split Node}
\usage{
getSplitNode(tree, taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{taxa}{Vector containing taxa to check}
}
\value{
Node ID of MRCA
}
\description{
This function returns the node ID of the MRCA of a set of 'taxa' in 'tree'
}
\examples{
getSplitNode(tree,taxa_to_check)

}
