% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSplits.R
\name{getSplits}
\alias{getSplits}
\title{Rboretum Tree Splitter}
\usage{
getSplits(tree)
}
\arguments{
\item{tree}{Phylo object}
}
\value{
Four-column dataframe; 1: Semi-colon separated monophyletic clade; 2: 'Mirror Clade'; 3: Phylo Node ID; 4: Node Boostrap (if present)
}
\description{
This function takes a phylo object and returns information about each monophyletic group/split
}
\examples{
getSplits(tree)

}
