% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeClades.R
\name{getTreeClades}
\alias{getTreeClades}
\title{Rboretum Clade Fetcher}
\usage{
getTreeClades(tree, include_root, print_counts, return_counts)
}
\arguments{
\item{tree}{Tree(s) to split. Options include:
\itemize{
  \item A rooted phylo object
  \item A rooted, named multiPhylo object where all trees share 3+ taxa
}}

\item{include_root}{OPTIONAL: If TRUE, return two root clades [Default: FALSE, exlclude root clades]}

\item{print_counts}{OPTIONAL: If TRUE, print a summary table of unique splits and how many/which trees contain them [Default: FALSE, no printing]}

\item{return_counts}{OPTIONAL: If TRUE, instead of returning a vector of all clades, return a summary table of unique splits and how many/which trees contain them [Default: FALSE, return clades as sorted vector]}
}
\value{
Dataframe (or a list of dataframes for each unique topology): Column 1: Semi-colon separated monophyletic clade; Column 2: 'Mirror Clade'; Column 3: Phylo Node ID (NA for root split)
}
\description{
This function breaks down a rooted phylo or multiPhylo object into its respective monophyletic clades (root clades excluded unless noted), sorted alphabetically.
}
