% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSplits.R
\name{getTreeSplits}
\alias{getTreeSplits}
\title{Rboretum Tree Splitter}
\usage{
getTreeSplits(tree)
}
\arguments{
\item{tree}{Tree(s) to split. Options include:
\itemize{
  \item A rooted phylo object
  \item A rooted multiPhylo object where all trees share 3+ taxa
}}
}
\value{
Dataframe (or a list of dataframes for each topology): Column 1: Semi-colon separated monophyletic clade; Column 2: 'Mirror Clade'; Column 3: Phylo Node ID (NA for root split)
}
\description{
This function breaks down a rooted phylo or multiPhylo object into its respective set of splits
}
