% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSplits_Worker.R
\name{getTreeSplits_Worker}
\alias{getTreeSplits_Worker}
\title{Rboretum Tree Splitter}
\usage{
getTreeSplits_Worker(tree)
}
\arguments{
\item{tree}{Rooted phylo object}
}
\value{
Dataframe: Column 1: Semi-colon separated monophyletic clade; Column 2: 'Mirror Clade'; Column 3: Phylo Node ID
}
\description{
This function breaks down a rooted phylo into its respective set of splits
}
