% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSupport.R
\name{getTreeSupport}
\alias{getTreeSupport}
\title{Rboretum Alignment Signal Support Mapper}
\usage{
getTreeSupport(
  tree,
  signal,
  max_missing,
  alignment_name,
  include_gap,
  include_singleton,
  include_biallelic,
  include_triallelic,
  include_quadallelic,
  include_pentallelic,
  only_gap,
  existing_support
)
}
\arguments{
\item{tree}{phylo object onto which alignment signal will be mapped}

\item{signal}{Output table from getAlignmentSignal() run with taxa from 'tree'}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column [Default: 0]}

\item{alignment_name}{OPTIONAL: Column name for data being added [Default: Alignment name from signal dataframe + 'm_<MISSING>']}

\item{include_gap}{OPTIONAL: TRUE or FALSE; Count sites with gap positions ('-') as part of total support [Default: TRUE]}

\item{include_singleton}{OPTIONAL: TRUE or FALSE; Count sites with singletons as part of total support [Default: TRUE]}

\item{include_biallelic}{OPTIONAL: TRUE or FALSE; Count sites with biiallelic variation as part of total support [Default: TRUE]}

\item{include_triallelic}{OPTIONAL: TRUE or FALSE; Count sites with triallelic variation as part of total support [Default: TRUE]}

\item{include_quadallelic}{OPTIONAL: TRUE or FALSE; Count sites with quadallelic variation as part of total support [Default: TRUE]}

\item{include_pentallelic}{OPTIONAL: TRUE or FALSE; Count sites with pentallelic variation as part of total support [Default: TRUE]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') as part of total support [Default: FALSE]}

\item{existing_support}{OPTIONAL: Output from previous getTreeSupport() using the same tree, run with new alignment/missing combination}
}
\value{
The same split table from getTreeSplits(tree), but with (1) no root row, and (2) a support column for the specfied alignment/missing combination
}
\description{
This function computes the alignment support for a specified tree
}
