% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSupport_Worker.R
\name{getTreeSupport_Worker}
\alias{getTreeSupport_Worker}
\title{Rboretum Alignment Signal Support Mapper}
\usage{
getTreeSupport_Worker(tree, signal, alignment_name)
}
\arguments{
\item{tree}{A rooted phylo object onto which alignment signal will be mapped.}

\item{signal}{Output table from getAlignmentSignal() run with taxa against 'tree'}

\item{alignment_name}{Column name(s) for data being added [Default: Alignment name from signal dataframe + 'm_<MISSING>']}
}
\value{
The same split table from getTreeSplits(tree), but with (1) no root row, and (2) a support column for the specfied alignment/missing combination
}
\description{
This function maps phylogenetic signal from a multiple-sequence alingment onto a rooted phylo or multiPhylo object
}
