% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrimmedTree.R
\name{getTrimmedTree}
\alias{getTrimmedTree}
\title{Get Trimmed Tree}
\usage{
getTrimmedTree(whole_tree, taxa)
}
\arguments{
\item{whole_tree}{Phylo object}

\item{taxa}{Vector containing desired taxa list}
}
\value{
Pruned phylo object
}
\description{
Given a tree and a list of taxa, this function returns a tree pruned to include only those taxa. Note: Function will STOP if all taxa are not in tree.
}
\examples{
getTrimmedTree(whole_tree,taxa)

}
