% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_add.new_aes.R
\name{ggplot_add.new_aes}
\alias{ggplot_add.new_aes}
\title{Add New Scale
Source: eliocamp/new_aes.R}
\usage{
\method{ggplot_add}{new_aes}(object, plot, object_name)
}
\description{
Add New Scale
Source: eliocamp/new_aes.R
}
