% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.multiPhylo.R
\name{is.multiPhylo}
\alias{is.multiPhylo}
\title{Rboretum MultiPhylo Checker}
\usage{
is.multiPhylo(test_object)
}
\arguments{
\item{test_object}{R object to check}
}
\value{
TRUE if multiPhylo with 2+ trees, otherwise FALSE
}
\description{
This function returns TRUE if the passed object is of class multiPhylo and has 2+ trees; Otherwise, FALSE
}
