% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.plot.R
\name{is.plot}
\alias{is.plot}
\title{Rboretum Plot Checker}
\usage{
is.plot(test_object)
}
\arguments{
\item{test_object}{R object to check}
}
\value{
TRUE if plot, otherwise FALSE
}
\description{
This function returns TRUE if the passed object is of class ggtree or ggplot; Otherwise, FALSE
}
