% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isAlignmentSignal.R
\name{isAlignmentSignal}
\alias{isAlignmentSignal}
\title{Rboretum Alignment Signal Checker}
\usage{
isAlignmentSignal(test_object, species_info)
}
\arguments{
\item{test_object}{R object to check}

\item{species_info}{OPTIONAL: Check that passed signal contains information about a specific set of taxa. Set of 3+ species that can be passed as either:
\itemize{
  \item phylo object from which species will be extracted; or
  \item multiPhylo object from which common species will be extracted; or
  \item Character vector of desired taxa
}}
}
\value{
TRUE if output of getAlignmentSignal(); otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is the output of getAlignmentSignal; otherwise, FALSE
}
