% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCladeSupport.R
\name{isCladeSupport}
\alias{isCladeSupport}
\title{Rboretum Clade Support Checker}
\usage{
isCladeSupport(test_object, tree)
}
\arguments{
\item{test_object}{R object to check}

\item{tree}{OPTIONAL: phylo object [Check that passed clade support contains information about a specific tree.]}
}
\value{
TRUE if output of getCladeSupport(); otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is the output of getCladeSupport; otherwise, FALSE
}
