% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPlot.R
\name{isPlot}
\alias{isPlot}
\title{Rboretum Plot Checker}
\usage{
isPlot(test_object)
}
\arguments{
\item{test_object}{R object to check}
}
\value{
TRUE if plot, otherwise FALSE
}
\description{
This function returns TRUE if the passed object is of class ggtree or ggplot; Otherwise, FALSE
}
