% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTreeSupport.R
\name{isTreeSupport}
\alias{isTreeSupport}
\title{Rboretum Tree Support Checker}
\usage{
isTreeSupport(test_object, tree, partial)
}
\arguments{
\item{test_object}{R object to check}

\item{tree}{Check that passed support object contains information about a specific set of trees. Options include:
\itemize{
  \item A rooted phylo object; or,
  \item A named, rooted multiPhylo object where all trees share 3+ taxa
}}

\item{partial}{OPTIONAL: Allow for parital matches (as long as all clades from 'tree' are present, return TRUE) [Default: FALSE, test all clades]}
}
\value{
TRUE if output of getTreeSupport(tree,...); otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is the output of getTreeSupport; otherwise, FALSE
}
