% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie_inset.R
\name{pie_inset}
\alias{pie_inset}
\title{pie_inset}
\usage{
pie_inset(
  tree_view,
  insets,
  width,
  height,
  hjust = 0,
  vjust = 0,
  x = "node",
  reverse_x = FALSE,
  reverse_y = FALSE
)
}
\arguments{
\item{tree_view}{tree view}
}
\value{
tree view with insets
}
\description{
add insets in a tree
}
\details{
From ggtree, but older version that doesn't scale insets based on plot size
}
\author{
Guangchuang Yu
}
