% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pies.treePlot.R
\name{pies.treePlot}
\alias{pies.treePlot}
\title{Rboretum Pie Graph Node Support Plotter}
\usage{
pies.treePlot(
  tree,
  tree_support,
  support_scales,
  node_alpha,
  legend_shape_size,
  legend_font_size,
  legend_title_size,
  legend_position,
  branch_length,
  branch_weight,
  node_label,
  node_size,
  node_nudge,
  pie_xnudge,
  pie_ynudge,
  taxa_size,
  taxa_italic,
  taxa_align,
  taxa_offset,
  xmax
)
}
\arguments{
\item{tree}{Rooted phylo object}

\item{tree_support}{OPTIONAL: Output trom tree.support() run on the same tree}

\item{support_scales}{OPTIONAL: Scaling factor for nodepoint labels. Options include:
\itemize{
  \item "log" - Log-converted support values
  \item Single numeric value: All nodepoint labels will be this size [Default: 3]
  \item c(x,y): Support values will be re-scaled from numeric values x-y
}}

\item{node_alpha}{OPTIONAL: ggplot2 alpha value for geom_nodepoint [Default: 0.9]}

\item{legend_shape_size}{OPTIONAL: ggplot2 size for legend icons [Default: 10]}

\item{legend_font_size}{OPTIONAL: ggplot2 size for legend font [Default: 10]}

\item{legend_title_size}{OPTIONAL: ggplot size for legend title [Default: 10]}

\item{legend_position}{OPTIONAL: Numerical vector of length four specifying legend xmin,xmax,ymin,ymax respectively. [Necessary oddity of generating a legend for nodepie data]}

\item{branch_length}{OPTIONAL: TRUE [plot tree with branch lengths]; FALSE [Default: plot cladogram]}

\item{branch_weight}{OPTIONAL: Set ggtree branch thickness}

\item{node_label}{OPTIONAL: Choice of node labels include:
\itemize{
  \item "none": No node labels [Default]
  \item "node": Node number
  \item "bs": Bootstrap value
  \item "support": Raw total support count
}}

\item{node_size}{OPTIONAL: Set ggtree node label size}

\item{node_nudge}{OPTIONAL: Set ggtree node label nudge_x}

\item{pie_xnudge}{OPTIONAL: Set ggtree pie label hjust}

\item{pie_ynudge}{OPTIONAL: Set ggtree pie label yjust}

\item{taxa_size}{OPTIONAL: Set ggtree tip label size}

\item{taxa_italic}{OPTIONAL: TRUE [italic tip labels]; FALSE [Default: non-italic tip labels]}

\item{taxa_align}{OPTIONAL: 'left' or 'right' tip label alignment [Default: No alignment]}

\item{taxa_offset}{OPTIONAL: Set ggtree tip label offset}

\item{xmax}{OPTIONAL: Set ggplot xlim upper limit (e.g if long tip labels run off plot)}
}
\value{
ggtree object
}
\description{
Given a phylo object (tree), and the output from tree.support(), this function plots the tree [adjusted with possible arguments] along with node support information
}
\examples{
# Print tree with bootstrap labels
basic.treePlot(tree)
}
