% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotShuffledSupport.R
\name{plotShuffledSupport}
\alias{plotShuffledSupport}
\title{Plot Alignment Support and Node Prevalence for Specific Topology}
\usage{
plotShuffledSupport(comparison, signal, tree, support_scales, xmax)
}
\arguments{
\item{comparison}{x}

\item{signal}{x}

\item{tree}{x}

\item{support_scales}{x}

\item{xmax}{x}
}
\value{
Tree with nodepoint geoms scaled to support, and colored by number of trees containing that split
}
\description{
DESCRIPTION
}
\examples{
plotShuffledSupport(comparison,signal,tree,support_scales,xmax)

}
