% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTreeSupportBreakdown.R
\name{plotTreeSupportBreakdown}
\alias{plotTreeSupportBreakdown}
\title{Plot Pie Charts for Support from Different Datasets}
\usage{
plotTreeSupportBreakdown(tree, tree_support, support_scales, xmax)
}
\arguments{
\item{tree}{x}

\item{tree_support}{x}

\item{support_scales}{x}

\item{xmax}{x}
}
\value{
x
}
\description{
DESCRIPTION
}
\examples{


}
