% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.rooted.R
\name{read.rooted}
\alias{read.rooted}
\title{Rboretum Rooted Tree Reader}
\usage{
treePath <- '/path/to/my/tree.nwk'
rootTaxa <- c('Root_species_1','Root_species_2')
rootedTree <- read.rooted(tree_path,root_taxa)
}
\arguments{
\item{tree_path}{Path to tree file that can be read by ape::read.tree() or ape::read.nexus()}

\item{root_taxa}{Character vector containing outgroup species IDs (Must be in tree and monophyletic)}

\item{resolve_root}{OPTIONAL: Set ape::resolve.root [Default: TRUE]}
}
\value{
A phylo object, rooted at specified taxa
}
\description{
This function is an ape wrapper, and returns a tree rooted as specified by the user.
}
\examples{
myRootedTetrapods <- read.rooted('/path/to/tetrapod_tree.nwk',c('Fish_1','Fish_2'))
}
