% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRootedPhylo.R
\name{readRootedPhylo}
\alias{readRootedPhylo}
\title{Read Rooted Phylogenetic Tree from Path}
\usage{
readRootedPhylo(tree_path, root_taxa)
}
\arguments{
\item{tree_path}{Path to tree file}

\item{root_taxa}{Vector containing outgroup species IDs (Must be in tree and monophyletic)}
}
\value{
A phylo object rooted at specified taxa
}
\description{
This function returns a rooted tree, assuming the tree contains the outgroup taxa, and they are monophyletic.
}
\examples{
readRootedPhylo('/path/to/tetrapod_tree.nwk',c('Fish_1','Fish_2'))
OR
tree_path <- '/path/to/tetrapod_tree.nwk'
outgroup_species <- c('Fish_1','Fish_2')
readRootedPhylo(tree_path,outgroup_species)
}
