% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRootedTree.R
\name{readRootedTree}
\alias{readRootedTree}
\title{Rboretum Rooted Tree Reader}
\usage{
readRootedTree(tree_path, root_taxa)
}
\arguments{
\item{tree_path}{Path to tree file that can be read by ape::read.tree() or ape::read.nexus()}

\item{root_taxa}{Character vector containing outgroup species IDs (Must be in tree and monophyletic)}
}
\value{
A phylo object, rooted at specified taxa
}
\description{
This function is an ape wrapper, and returns a tree rooted as specified by the user.
}
