% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerootTree.R
\name{rerootTree}
\alias{rerootTree}
\title{Re-Root Phylogenetic Tree}
\usage{
rerootTree(tree, root_taxa)
}
\arguments{
\item{tree}{Path to tree file}

\item{root_taxa}{Vector containing outgroup species IDs (Must be in tree and monophyletic)}
}
\value{
A phylo object rooted at specified taxa
}
\description{
This function returns a rooted tree, assuming the tree contains the outgroup taxa, and they are monophyletic.
}
\examples{
rerootTree(tree,taxa)
}
