% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.taxa.R
\name{same.taxa}
\alias{same.taxa}
\title{Rboretum Identical Taxon Checker}
\usage{
same.taxa(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
TRUE if all trees have all species in common; else, FALSE
}
\description{
This function takes a multiPhylo object and returns TRUE if all species are shared among all trees; otherwise, FALSE
}
\examples{
trees <- c(tree_1,tree_2,tree_3,...)
same.taxa(trees)

}
