% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/same.topology.R
\name{same.topology}
\alias{same.topology}
\title{Rboretum Topology Checker}
\usage{
same.topology(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
TRUE if all trees have the same topology after pruning to identical species lists; else, FALSE
}
\description{
This function returns TRUE if all trees in multiPhylo have the same topology after pruning to identical species lists
}
\examples{
myTrees <- c(Tree_1,Tree_2,Tree_3)
same.topology(trees)

}
