% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiChecker.R
\name{semiChecker}
\alias{semiChecker}
\title{Rboretum Semi Checker}
\usage{
semiChecker(test_object, return_count)
}
\arguments{
\item{test_object}{Object to check}

\item{return_count}{OPTIONAL: If TRUE and test_object is a semicolon-separated character, return the number of elements after splitting [Default: FALSE]}
}
\value{
TRUE if test_object is a semicolon-separated character (or element count if return_count = TRUE), or FALSE
}
\description{
This function returns TRUE if the passed object is a semicolon-separated character with a length = 1. If return_count = TRUE, it will return the number of elements after splitting
}
