% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiSorter.R
\name{semiSorter}
\alias{semiSorter}
\title{Rboretum Semicolon Sorter}
\usage{
semiSorter(string_to_sort)
}
\arguments{
\item{string_to_split}{Character vector or unsorted semicolon delimited string}
}
\value{
Sorted semicolon delimited string
}
\description{
This function returns a sorted, semicolon-delimited string from a character vector or an unsorted semicolon-delimited string
}
\examples{
myShuffledString <- 'c;d;a;b'
semiSorter(myShuffledString)
> 'a;b;c;d'
}
