% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sharedNodePlotter.R
\name{sharedNodePlotter}
\alias{sharedNodePlotter}
\title{Basic Tree Plotter}
\usage{
sharedNodePlotter(tree, split_df)
}
\arguments{
\item{tree_1}{Phylo object}

\item{label}{Either 'node' [print node labels] or 'bootstrap' [print bootstrap labels]. Any other entry (or missing) results in no node labels}
}
\description{
This function plots a tree without branch lengths, and nodes can be labeled with bootstrap or node labels
}
\examples{
# Print tree with no labels
basicTreePlot(tree)

# Print tree with bootstrap labels
basicTreePlot(tree,'bootstrap')

# Print tree with node labels
basicTreePlot(tree,'node')

}
