% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.semiVector.R
\name{simple.semiVector}
\alias{simple.semiVector}
\title{Rboretum Semicolon Vectorizer}
\usage{
semiVector(string_to_split)
}
\arguments{
\item{string_to_split}{Semicolon delimited string}
}
\value{
Character vector
}
\description{
This function returns a vector of items from a string separated by semicolons (;)
}
\examples{
myString <- 'a;b;c;d'
semiVector(myString)
> ['a','b','c','d']

}
