% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripNodeLabels.R
\name{stripNodeLabels}
\alias{stripNodeLabels}
\title{Rboretum Node Label Stripper}
\usage{
stripNodeLabels(tree)
}
\arguments{
\item{tree}{A phylo or multiPhylo object}
}
\value{
A phylo or multiPhylo object with no node labels
}
\description{
This function takes a phylo or multiPhylo of trees and removed node labels
}
