% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMultiPhylo.R
\name{summarizeMultiPhylo}
\alias{summarizeMultiPhylo}
\title{Rboretum MultiPhylo Summarizer}
\usage{
summarizeMultiPhylo(trees)
}
\arguments{
\item{trees}{A rooted multiPhylo object where all trees share 3+ taxa}
}
\description{
This function breaks down a rooted multiPhylo object into a number of summary values
}
