% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.treePlot.R
\name{support.treePlot}
\alias{support.treePlot}
\title{Rboretum Site and Node Support Plotter}
\usage{
support.treePlot(
  tree,
  tree_support,
  support_scales,
  clade_support,
  branch_length,
  branch_weight,
  node_label,
  node_size,
  node_nudge,
  taxa_size,
  taxa_italic,
  taxa_align,
  taxa_offset,
  xmax
)
}
\arguments{
\item{tree}{Rooted phylo object}

\item{tree_support}{OPTIONAL: Output trom tree.support() run on the same tree}

\item{support_scales}{OPTIONAL: Scaling factor for nodepoint labels. Options include:
\itemize{
  \item "log" - Log-converted support values
  \item Single numeric value: All nodepoint labels will be this size [Default: 1]
  \item c(x,y): Support values will be re-scaled from numeric values x-y
}}

\item{clade_support}{OPTIONAL: Output from compare.clades(). Will colorize nodepoint labels based on how many trees in multiPhylo contain that split}

\item{branch_length}{OPTIONAL: TRUE [plot tree with branch lengths]; FALSE [Default: plot cladogram]}

\item{branch_weight}{OPTIONAL: Set ggtree branch thickness}

\item{node_label}{OPTIONAL: Choice of node labels include:
\itemize{
  \item "none": No node labels [Default]
  \item "node": Node number
  \item "bs": Bootstrap value
  \item "support": Raw total support count
}}

\item{node_size}{OPTIONAL: Set ggtree node label size}

\item{node_nudge}{OPTIONAL: Set ggtree node label nudge_x}

\item{taxa_size}{OPTIONAL: Set ggtree tip label size}

\item{taxa_italic}{OPTIONAL: TRUE [italic tip labels]; FALSE [Default: non-italic tip labels]}

\item{taxa_align}{OPTIONAL: 'left' or 'right' tip label alignment [Default: No alignment]}

\item{taxa_offset}{OPTIONAL: Set ggtree tip label offset}

\item{xmax}{OPTIONAL: Set ggplot xlim upper limit (e.g if long tip labels run off plot)}
}
\value{
ggtree object
}
\description{
Given a phylo object (tree), and the output from tree.support(), this function plots the tree [adjusted with possible arguments] along with node support information
}
\examples{
# Print tree with bootstrap labels
basic.treePlot(tree)
}
