% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportPlot.R
\name{supportPlot}
\alias{supportPlot}
\title{Basic Tree Plotter}
\usage{
supportPlot(tree, tree_support, support_scales, xmax)
}
\arguments{
\item{tree}{Phylo object}

\item{tree_support}{Output from getTreeSupport}

\item{support_scales}{OPTIONAL: Vector of values to rescale support geoms}
}
\description{
This function plots a tree without branch lengths, and nodes can be labeled with bootstrap or node labels
}
\examples{
# Set up data
tree< - myTree
signal <- getAlignmentSignal(alignment_path,species_info = myTree)
tree_support <- getTreeSupport(tree,signal)

supportPlot(tree,tree_support)

}
