% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCount.R
\name{tableCount}
\alias{tableCount}
\title{Rboretum Table Search}
\usage{
tableCount(search_table, name)
}
\arguments{
\item{search_table}{R Table (Result of 'table()' call)}

\item{name}{Name of item you want the count for}
}
\value{
Number of times that value occurs in the table
}
\description{
This function returns the count of items in an R table, or 0 if item does not occur in table
}
