% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.treePlot.R
\name{tandem.treePlot}
\alias{tandem.treePlot}
\title{Rboretum Simple Tandem Plotter
Simple cowplot wrapper to plot two or more ggtree/ggplot objects side by side}
\usage{
tandem.treePlot(...)
}
\value{
Side-by-side ggplot object
}
\description{
Rboretum Simple Tandem Plotter
Simple cowplot wrapper to plot two or more ggtree/ggplot objects side by side
}
\examples{
tandem.treePlot(ggtree1,ggtree2)
}
