% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timePlotter.R
\name{timePlotter}
\alias{timePlotter}
\title{Rboretum Time-Dependent Signal Plotter}
\usage{
timePlotter(
  node_age_df,
  tree_support,
  all_sites_col,
  lm_alpha,
  wrap,
  wrap_scales,
  return_stats
)
}
\arguments{
\item{node_age_df}{}

\item{tree_support}{}

\item{all_sites_col}{}

\item{lm_alpha}{}
}
\value{

}
\description{
This function takes information about node ages and node support, and plots relative changes in node support over time among datasets.
}
