% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipCheck.R
\name{tipCheck}
\alias{tipCheck}
\title{Rboreturm Tip Checker}
\usage{
tipCheck(tree, taxa)
}
\arguments{
\item{tree}{phylo or multiPhylo object}

\item{taxa}{Character vector containing taxa to query}
}
\value{
TRUE if all 'taxa' in 'tree'; else, FALSE
}
\description{
This function returns TRUE if all 'taxa' are present in 'tree' (or a multiPhylo of trees); FALSE otherwise
}
