% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.support.R
\name{tree.support}
\alias{tree.support}
\title{Rboretum Alignment Signal Support Mapper}
\usage{
tree.support(signal, tree, max_missing, existing_splits)
}
\arguments{
\item{signal}{Output table from alignment.signal()}

\item{tree}{Rooted phylo object [Note: Taxa must EXACTLY match those passed to alignment.signal()]}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column [Default: 0]}

\item{existing_splits}{OPTIONAL: Output from previous tree.support() using the same tree, run with new alignment/missing combination}
}
\value{
The same split table from get.splits(tree), but with a support column for the specfied alignment/missing combination
}
\description{
This function computes the alignment support for a specified tree, processing signal from all alignment columns with <= 'max_missing' missing taxa [from output of alignment.signal()]. Signal from multiple alignments/missing combinations can be added to the same table by passing the output of tree.support() to the 'existing_splits' argument
}
\examples{
tree.support(tree,signal)

}
