% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeNamer.R
\name{treeNamer}
\alias{treeNamer}
\title{Rboretum MultiPhylo Autonamer}
\usage{
treeNamer(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
A named multiPhylo object
}
\description{
This function adds/replaces names of multiPhylo trees with dummy names (Tree_1, Tree_2, etc.)
}
