% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlotter.R
\name{treePlotter}
\alias{treePlotter}
\title{Rboretum Tree Plotter}
\usage{
treePlotter(
  tree,
  branch_length,
  branch_weight,
  node_label,
  node_size,
  node_nudge,
  taxa_size,
  taxa_italic,
  taxa_align,
  taxa_offset,
  xmax
)
}
\arguments{
\item{tree}{Phylo object}

\item{branch_length}{OPTIONAL: TRUE or FALSE [Default: FALSE (print cladogram)]}

\item{branch_weight}{OPTIONAL: Override default ggtree branch thickness}

\item{node_label}{OPTIONAL: 'none', 'node', or 'bs' [Default: bs (print bootstrap)]}

\item{node_size}{OPTIONAL: Override default ggtree node label size}

\item{node_nudge}{OPTIONAL: Override default ggtree node label placement}

\item{taxa_size}{OPTIONAL: Override default ggtree tip label size}

\item{taxa_italic}{OPTIONAL: TRUE or FALSE [Default: FALSE]}

\item{taxa_align}{OPTIONAL: 'left' or 'right' tip label alignment [Default: No alignment]}

\item{taxa_offset}{OPTIONAL: Override default ggtree tip label placement}

\item{xmax}{OPTIONAL: Override default ggplot xlim upper limit (e.g + xlim(0,max))}
}
\description{
Given a phylo object (tree), this ggtree wrapper plots the tree with following optional arguments:
- node_label: 'none' [no node labels], 'node' [print node number labels] or 'bs' [print bootstrap labels] (Default: Plot with bootstrap labels, if present)
- node_size: Set ggtree node label size (Default: ggtree default)
- node_nudge: Set ggtree nudge_x for node label
- branch_length: TRUE or FALSE; Plot tree with or without branch lengths (Default: FALSE; Plot as cladogram) 
- branch_weight: Set  ggtree branch thickness (Default: ggTree default)
- taxa_size: Set ggtree tip label size (Default: ggTree default)
- taxa_italic: TRUE of FALSE; Print tip labels in italic (Default: False)
- taxa_align: 'left' or 'right' alignment of tip labels
- taxa_offset: Set ggtree offset tip label value (Default: No offset)
- xmax: numeric extension value for ggplot2::xlim(0,xmax) (e.g. if tip labels are cut-off)
}
\examples{
# Print tree with bootstrap labels
treePlotter(tree)

# Print tree with no node labels
treePlotter(tree,node_label='none')

# Print tree with thicker branches
treePlotter(tree,branch_weight=3)

# Print tree with buffer on right side to accomodate longer tip labels
treePlotter(tree,xmax=10)
}
