% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeSupport.R
\name{treeSupport}
\alias{treeSupport}
\title{Rboretum Alignment Signal Support Mapper}
\usage{
treeSupport(
  signal,
  tree,
  max_missing,
  alignment_name,
  include_gap,
  include_singleton,
  include_biallelic,
  include_triallelic,
  include_quadallelic,
  include_pentallelic,
  only_gap,
  existing_splits
)
}
\arguments{
\item{signal}{Output table from getAlignmentSignal()}

\item{tree}{Rooted phylo object [Note: Taxa must EXACTLY match those passed to getAlignmentSignal()]}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column [Default: 0]}

\item{alignment_name}{OPTIONAL: Column name for data being added [Default: Alignment name from signal dataframe]}

\item{include_gap}{OPTIONAL: TRUE or FALSE; Count sites with gap positions ('-') as part of total support [Default: TRUE]}

\item{include_singleton}{OPTIONAL: TRUE or FALSE; Count sites with singletons as part of total support [Default: TRUE]}

\item{include_biallelic}{OPTIONAL: TRUE or FALSE; Count sites with biiallelic variation as part of total support [Default: TRUE]}

\item{include_triallelic}{OPTIONAL: TRUE or FALSE; Count sites with triallelic variation as part of total support [Default: TRUE]}

\item{include_quadallelic}{OPTIONAL: TRUE or FALSE; Count sites with quadallelic variation as part of total support [Default: TRUE]}

\item{include_pentallelic}{OPTIONAL: TRUE or FALSE; Count sites with pentallelic variation as part of total support [Default: TRUE]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') as part of total support [Default: FALSE]}

\item{existing_splits}{OPTIONAL: Output from previous tree.support() using the same tree, run with new alignment/missing combination}
}
\value{
The same split table from getTreeSplits(tree), but with a support column for the specfied alignment/missing combination
}
\description{
This function computes the alignment support for a specified tree, processing signal from all alignment columns with <= 'max_missing' missing taxa [from output of getAlignmentSignal()]. Signal from multiple alignments/missing combinations can be added to the same table by passing the output of tree.support() to the 'existing_splits' argument
}
