% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeTrimmer.R
\name{treeTrimmer}
\alias{treeTrimmer}
\title{Rboretum Tree Trimmer}
\usage{
treeTrimmer(tree, taxa, remove)
}
\arguments{
\item{tree}{phylo or multiPhylo object}

\item{taxa}{Character vector of desired tip labels to keep (or discard if remove=TRUE)}

\item{remove}{OPTIONAL: If TRUE, tip labels specified by 'taxa' are removed from all trees rather than retained [Default: FALSE, prune 'tree' down to 'taxa']}
}
\value{
Pruned phylo or multiPhylo object
}
\description{
This ape wrapper returns a phylo or multiPhylo object that has been pruned to include only specified taxa. Note: Function will STOP if any 'tree' is missing any 'taxa'.
}
