% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorSemi.R
\name{vectorSemi}
\alias{vectorSemi}
\title{Rboretum Vector Semicolonizer}
\usage{
vectorSemi(char_vec)
}
\arguments{
\item{char_vec}{Character vector}
}
\value{
Semicolon-separated character vector of length 1
}
\description{
This function takes a character vector and retuns a string separated by semicolons (;)
}
\examples{
myVec <- c('a','b','c','d')
vectorSemi(myVec)
> 'a;b;c;d'
}
