% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_getTreeSupport.R
\name{batch_getTreeSupport}
\alias{batch_getTreeSupport}
\title{Rboretum Batch Alignment Signal Support Mapper}
\usage{
batch_getTreeSupport(
  signal,
  tree,
  max_missing,
  alignment_name,
  include_gap,
  include_singleton,
  include_biallelic,
  include_triallelic,
  include_quadallelic,
  include_pentallelic,
  only_gap,
  existing_splits
)
}
\arguments{
\item{signal}{\itemize{
  \item Output table from getAlignmentSignal(); or, 
  \item Output table from batch_getAlignmentSignal(); or, 
}}

\item{tree}{Tree(s) on which to map alignment support. Options include:
\itemize{
  \item A single phylo object where taxa EXACTLY match those from signal [Returns results as a single dataframe with multiple support columns]; or, 
  \item A named multiPhylo object, where all trees have (1) taxa that EXACTLY match those from signal and (2) a unique topology [Returns results as a named list of dataframes]
}}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column [Default: 0]}

\item{alignment_name}{OPTIONAL: Column name for data being added [Default: Alignment name from signal dataframe]}

\item{include_gap}{OPTIONAL: TRUE or FALSE; Count sites with gap positions ('-') as part of total support [Default: TRUE]}

\item{include_singleton}{OPTIONAL: TRUE or FALSE; Count sites with singletons as part of total support [Default: TRUE]}

\item{include_biallelic}{OPTIONAL: TRUE or FALSE; Count sites with biiallelic variation as part of total support [Default: TRUE]}

\item{include_triallelic}{OPTIONAL: TRUE or FALSE; Count sites with triallelic variation as part of total support [Default: TRUE]}

\item{include_quadallelic}{OPTIONAL: TRUE or FALSE; Count sites with quadallelic variation as part of total support [Default: TRUE]}

\item{include_pentallelic}{OPTIONAL: TRUE or FALSE; Count sites with pentallelic variation as part of total support [Default: TRUE]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') as part of total support [Default: FALSE]}

\item{existing_splits}{OPTIONAL: Output from previous tree.support() using the same tree, run with new alignment/missing combination}
}
\value{
The same split table from getTreeSplits(tree), but with a support column for the specfied alignment/missing combination
}
\description{
This function computes the support for a specified tree or a set of trees from an alignment or set of alignments
}
