% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkComparable.R
\name{checkComparable}
\alias{checkComparable}
\title{Check if Pair of Tree Topologies Can Be Compared}
\usage{
checkComparable(tree_1, tree_2)
}
\arguments{
\item{tree_1}{Phylo object}

\item{tree_2}{Phylo object}
}
\value{
TRUE if 'tree_1' and 'tree_2' have 3 or more species in commmon and a unique topology; else, FALSE
}
\description{
This function returns TRUE if 'tree_1' and 'tree_2' have 3 or more species in commmon, and a unique topology
}
\examples{
checkComparable(tree_1,tree_2)

}
