% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIdenticalSpecies.R
\name{checkIdenticalSpecies}
\alias{checkIdenticalSpecies}
\title{Rboretum Identical Taxon Checker}
\usage{
checkIdenticalSpecies(trees)
}
\arguments{
\item{trees}{Multiphylo object}
}
\value{
TRUE if all trees have all species in common; else, FALSE
}
\description{
This function takes a multiphylo object and returns TRUE if all species are shared among all trees; otherwise, FALSE
}
\examples{
trees <- c(tree_1,tree_2,tree_3,...)
checkIdenticalSpeciesMulti(trees)

}
