% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkSharedSpecies.R
\name{checkSharedSpecies}
\alias{checkSharedSpecies}
\title{Rboretum Three Common Taxa Checker}
\usage{
checkSharedSpecies(trees)
}
\arguments{
\item{trees}{Multiphylo object}
}
\value{
TRUE (all trees share >= 3 species) or FALSE
}
\description{
This function takes a multiphylo object and returns TRUE if all trees share at least three common taxa (e.g. enough to make a tree); otherwise, FALSE
}
\examples{
trees <- c(tree_1,tree_2,tree_3)
checkSharedSpeciesMulti(trees)

}
