% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTips.R
\name{checkTips}
\alias{checkTips}
\title{Rboreturm Tip Checker (X)}
\usage{
checkTips(tree, taxa, check_mono, check_root)
}
\arguments{
\item{tree}{Tree(s) to check. Options include:
\itemize{
  \item A phylo object
  \item A multiPhylo object where all trees share 3+ taxa
}}

\item{taxa}{Taxon labels to check. Can be:
\itemize{
  \item A character vector of tip labels
  \item A semicolon-separated list of labels
}}

\item{check_mono}{OPTIONAL: If TRUE, also query whether 'taxa' are monophyletic in tree(s) [Default: FALSE]}

\item{check_root}{OPTIONAL: If TRUE, also query whether 'taxa' are monophyletic and make up one of the two root splits in tree(s) [Default: FALSE]}
}
\value{
TRUE if all 'taxa' in 'tree'; else, FALSE (barring other selected options)
}
\description{
This function returns TRUE if all 'taxa' are present in 'tree' (or a multiPhylo of trees); FALSE otherwise. If multiPhylo, all comparisons are made on the common set of taxa.
}
\examples{
# Check if all labels in 'myTaxa' are present in phylo object 'myTree'
myTaxa <- c('Species_1','Species_2','Species_3')
checkTips(myTree,myTaxa)
# Check if all labels in 'myTaxa' are present in all trees in multiPhylo object 'myTrees'
checkTips(myTrees,myTaxa)
# Check if clade 'myTaxa' are monophyletic in phylo object 'myTree'
checkTips(myTree,myTaxa,check_mono=TRUE)
# Check if clade 'myTaxa' are one of the two root splits in phylo object 'myTree' 
checkTips(myTree,myTaxa,check_root=TRUE)
}
