% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTreeTaxa.R
\name{checkTreeTaxa}
\alias{checkTreeTaxa}
\title{Rboreturm Phylo Taxon Checker}
\usage{
checkTreeTaxa(tree, taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{taxa}{Character vector containing taxa to query}
}
\value{
TRUE if all 'taxa' in 'tree', else, FALSE
}
\description{
This function returns TRUE if all 'taxa' are present in 'tree', FALSE otherwise
}
\examples{

taxa_to_find <- c('Spp1','Spp2','Spp3')
checkTreeTaxa(myTree,taxa_to_find)

}
