% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkTreeTaxaMulti.R
\name{checkTreeTaxaMulti}
\alias{checkTreeTaxaMulti}
\title{Check Tree Taxa in Multiphylo}
\usage{
checkTreeTaxaMulti(trees, taxa)
}
\arguments{
\item{trees}{Multiphylo object}

\item{taxa}{Vector containing taxa to check}
}
\value{
TRUE if 'taxa' in all trees, else, FALSE
}
\description{
This function returns TRUE if all 'taxa' are present in all trees in a multiphylo, FALSE otherwise
}
\examples{
trees <- c(tree_1,tree_2,tree_3)
checkTreeTaxaMulti(trees,taxa)

}
