% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTipLabels.R
\name{convertTipLabels}
\alias{convertTipLabels}
\title{Rboretum Tip Label Converter}
\usage{
convertTipLabels(tree, name_df, from, to)
}
\arguments{
\item{tree}{phylo object}

\item{name_df}{Dataframe with name equivalencies}

\item{from}{Column name with current tree tip IDs (if missing, first column of name_df used as default)}

\item{to}{Column name with desired IDs (if missing, second column of name_df used as default)}
}
\value{
phylo object with converted tip labels
}
\description{
This function takes a tree and changes tip labels based on a supplied dataframe
}
