% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentSignal.R
\name{getAlignmentSignal}
\alias{getAlignmentSignal}
\title{Rboretum Alignment Signal Fetcher}
\usage{
getAlignmentSignal(
  alignment_path,
  species_info,
  use_gaps,
  alignment_name,
  prefix,
  suffix,
  existing_signal
)
}
\arguments{
\item{alignment_path}{Where to find alignment files. Options include:
\itemize{
  \item A character vector of one or more alignment file paths  (relative or absolute)
  \item A path to a single directory containing all alignment files (relative or absolute)
}}

\item{species_info}{OPTIONAL: List of taxa to analyze [Default: Process alignment using all taxa if solo, all shared in multiple]. Can be provided as:
\itemize{
  \item phylo object from which species will be extracted; or
  \item multiPhylo object from which common species will be extracted; or
  \item Character vector with 3+ taxon IDs
  \item Semicolon-separated list of taxon IDs
}}

\item{use_gaps}{OPTIONAL: If FALSE, treat gaps as missing data (like N) [Default: TRUE, treat gaps as indel data]}

\item{alignment_name}{OPTIONAL: Chacter vector of names for each alignment. If missing or incomplete, the base filename is used}

\item{prefix}{OPTIONAL: If 'alignment_path' is a directory, provide a character vector of file prefixes (e.g. all alignment files start with "Mafft_")}

\item{suffix}{OPTIONAL: If 'alignment_path' is a directory, provide a character vector of file suffixes (e.g. all alignment files end with ".phy")}

\item{existing_signal}{OPTIONAL: Append these results to the output from getAlignmentSignal() run with the same species_info and a different alignment}
}
\value{
Dataframe containing split and base information for each site in the alignment
}
\description{
Given the path(s) to one or more alignments (and an optional list of taxa), this script returns split and base information for each site
}
