% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAlignmentSignal.R
\name{getAlignmentSignal}
\alias{getAlignmentSignal}
\title{Rboretum Alignment Signal Fetcher}
\usage{
getAlignmentSignal(
  alignment_path,
  species_info,
  use_gaps,
  alignment_name,
  prefix,
  suffix
)
}
\arguments{
\item{alignment_path}{Where to find alignment files. Options include:
\itemize{
  \item A character vector of one or more alignment file paths  (relative or absolute)
  \item A path to a single directory containing all alignment files (relative or absolute)
}}

\item{species_info}{List of taxa to analyze. Can be provided as:
\itemize{
  \item phylo object from which species will be extracted; or
  \item Character vector with 3+ taxon IDs
}}

\item{use_gaps}{OPTIONAL: Options include:
\itemize{
  \item FALSE (Default: Treat all gaps (-) in all alignments as missing data)
  \item TRUE (Treat all gaps in all alignments as potentially informative signal)
  \item Logical vecor indicating TRUE or FALSE for each alignment
}}

\item{alignment_name}{Chacter vector of names for each alignment. If missing or incomplete, the base filename is used}

\item{prefix}{OPTIONAL: If 'alignment_path' is a directory, provide a character vector of file prefixes (e.g. all alignment files start with "Mafft_")}

\item{suffix}{OPTIONAL: If 'alignment_path' is a directory, provide a character vector of file suffixes (e.g. all alignment files end with ".phy")}
}
\value{
Dataframe containing split pattern for each site, in each alignment, relative to the given set of taxa
}
\description{
Given the path(s) to multiple alignments and a common list of taxa, this script returns site patterns for each site, in each alignment (after pruning if needed)
}
