% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllClades.R
\name{getAllClades}
\alias{getAllClades}
\title{Rboretum Clade Fetcher}
\usage{
getAllClades(tree)
}
\arguments{
\item{tree}{Phylo object}
}
\value{
Character vector of semicolon-separated monophyletic clades
}
\description{
This function takes a tree and returns a sorted character vector containing each monophyletic group
}
\examples{
getAllClades(tree)

}
