% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllClades.R
\name{getAllClades}
\alias{getAllClades}
\title{Get All Clades}
\usage{
getAllClades(rooted_tree, exclude_root)
}
\arguments{
\item{rooted_tree}{Rooted phylo object}

\item{exclude_root}{OPTIONAL: If TRUE, exclude root clades from returned list. Otherwise, return all groups (Default: FALSE)}
}
\value{
Character vector of monophyletic clades
}
\description{
This function takes a rooted tree and returns a vector containing each monophyletic group, including both root groups
}
\examples{
getAllClades(rooted_tree)

}
