% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSharedSpecies.R
\name{getSharedSpecies}
\alias{getSharedSpecies}
\title{Rboretum Common Taxa Fetcher}
\usage{
getSharedSpecies(trees)
}
\arguments{
\item{trees}{Multiphylo object}
}
\value{
Sorted character vector of tip labels present in all trees
}
\description{
This function takes a multiphylo object and returns a sorted list of taxa common to all trees
}
\examples{
trees <- c(tree_1,tree_2,tree_3)
getSharedSpeciesMulti(trees)

}
