% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeClades.R
\name{getTreeClades}
\alias{getTreeClades}
\title{Rboretum Clade Fetcher}
\usage{
getTreeClades(tree)
}
\arguments{
\item{tree}{Rooted phylo object}
}
\value{
Character vector of semicolon-separated monophyletic clades
}
\description{
This function takes a tree and returns a sorted character vector containing each monophyletic group (not including the root clades)
}
