% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeClades.R
\name{getTreeClades}
\alias{getTreeClades}
\title{Rboretum Clade Fetcher}
\usage{
getTreeClades(tree, include_root, print_counts, return_counts, return_shared)
}
\arguments{
\item{tree}{Tree(s) to split. Options include:
\itemize{
  \item A rooted phylo object
  \item A rooted multiPhylo object where all trees share 3+ taxa
}}

\item{include_root}{OPTIONAL: If TRUE, include both root clades (Disabled if tree is rooted on a single taxon) [Default: FALSE, exlclude root clades]}

\item{print_counts}{OPTIONAL: If TRUE, print a summary table of unique splits and how many/which trees contain them [Default: FALSE, no printing]}

\item{return_counts}{OPTIONAL: If TRUE, instead of returning a vector of clades, return a summary table of unique splits and how many/which trees contain them [Default: FALSE, return clades as sorted vector]}

\item{return_shared}{OPTIONAL: If TRUE, instead of returning a vector of clades, return a vector of only those clades present in all trees [Default: FALSE, return all clades]}
}
\value{
Either:
\itemize{
  \item A character vector of monophyletic clades; or,
  \item A dataframe containing information about clades and their occurrence among trees
}
}
\description{
This function breaks down a rooted phylo or multiPhylo object into its respective monophyletic clades (root clades excluded unless noted), sorted alphanumerically.
}
