% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSupport.R
\name{getTreeSupport}
\alias{getTreeSupport}
\title{Rboretum Alignment Signal Support Mapper}
\usage{
getTreeSupport(
  signal,
  tree,
  include_root,
  clade,
  dataset_name,
  max_missing,
  separate_signal,
  only_parsinf,
  include_gap,
  only_gap,
  only_biallelic,
  only_triallelic,
  only_quadallelic,
  only_pentallelic,
  return_integer,
  return_table,
  existing_support
)
}
\arguments{
\item{signal}{Output table from getAlignmentSignal()}

\item{tree}{OPTIONAL: Tree(s) onto which alignment signal will be mapped. Not considered if 'clade' argument is used. Tree options include:
\itemize{
  \item A rooted phylo object
  \item A rooted multiPhylo object where all trees share 3+ taxa
}}

\item{include_root}{OPTIONAL: If TRUE and using a 'tree',  return alignment support for root clades as well [Default: FALSE, don't include root clades]}

\item{clade}{OPTIONAL: Character vector of semicolon-separted taxa specifying specific clades of interest (i.e. get support for specific clade(s) rather than a whole tree). Supercedes 'tree' argument}

\item{dataset_name}{OPTIONAL: Character vector containing a new name for each alignment dataset in 'signal',  [Default: Alignment name from signal dataframe + 'm_<MISSING>']}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column before it is not considered [Default: Taxa Count - 3]}

\item{separate_signal}{OPTIONAL: If FALSE, return values as the sum of all datasets [Default: TRUE, return results separated by dataset]}

\item{only_parsinf}{OPTIONAL: If TRUE, do not count sites with singletons as part of total support (i.e. parsimony-informative sites) [Default: FALSE, allow sites with singletons to contribute data]}

\item{include_gap}{OPTIONAL: If FALSE, count sites with gap positions ('-') as missing data; otherwise, count gaps as valid indel data [Default: TRUE: Gaps are treated as indel signal]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') [Default: FALSE]}

\item{only_biallelic}{OPTIONAL: If TRUE, only count sites with biiallelic variation as part of total support [Default: FALSE]}

\item{only_triallelic}{OPTIONAL: If TRUE, only count sites with triallelic variation as part of total support [Default: FALSE]}

\item{only_quadallelic}{OPTIONAL: If TRUE, only count sites with quadallelic variation as part of total support [Default: FALSE]}

\item{only_pentallelic}{OPTIONAL: If TRUE, only count sites with pentallelic variation as part of total support [Default: FALSE]}

\item{return_integer}{OPTIONAL: If TRUE, return the integer support summed across datasets [Default: FALSE, return results as a dataframe]}

\item{return_table}{OPTIONAL: If TRUE, return entire table of filtered signal counts by clade [FALSE: Return clade support counts]}

\item{existing_support}{OPTIONAL: Append these results to the output from getTreeSupport() run with the same 'tree' and different alignment options}
}
\value{
A dataframe containing each monophyletic clade in 'tree', along with site support from all alignments in 'signal' as separate columns
}
\description{
This function maps phylogenetic signal from a multiple-sequence alingment onto a rooted phylo or multiPhylo object, or a set of specified clades
}
