% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSupport.R
\name{getTreeSupport}
\alias{getTreeSupport}
\title{Rboretum Alignment Signal Support Mapper}
\usage{
getTreeSupport(
  tree,
  signal,
  alignment_name,
  max_missing,
  include_gap,
  only_gap,
  include_singleton,
  include_biallelic,
  include_triallelic,
  include_quadallelic,
  include_pentallelic
)
}
\arguments{
\item{tree}{Tree(s) onto which alignment signal will be mapped. Options include:
\itemize{
  \item A rooted phylo object
  \item A rooted, named multiPhylo object where all trees share 3+ taxa
}}

\item{signal}{Output table from getAlignmentSignal() run with taxa against 'tree'}

\item{alignment_name}{OPTIONAL: Column name(s) for data being added [Default: Alignment name from signal dataframe + 'm_<MISSING>']}

\item{max_missing}{OPTIONAL: Number of missing sites allowed in alignment column before it is not considered [Default: 0, no missing taxa allowed]}

\item{include_gap}{OPTIONAL: If TRUE, count sites with gap positions ('-') as informative signal; otherwise, count gaps as missing data [Default: FALSE: Gaps are treated as missing]}

\item{only_gap}{OPTIONAL: TRUE or FALSE; Only count sites with gap positions ('-') [Default: FALSE]}

\item{include_singleton}{OPTIONAL: If TRUE, count sites with singletons as part of total support [Default: TRUE]}

\item{include_biallelic}{OPTIONAL: If TRUE, count sites with biiallelic variation as part of total support [Default: TRUE]}

\item{include_triallelic}{OPTIONAL: If TRUE, count sites with triallelic variation as part of total support [Default: TRUE]}

\item{include_quadallelic}{OPTIONAL: If TRUE, count sites with quadallelic variation as part of total support [Default: TRUE]}

\item{include_pentallelic}{OPTIONAL: If TRUE, count sites with pentallelic variation as part of total support [Default: TRUE]}

\item{existing_support}{OPTIONAL: Append these results to the output from getTreeSupport() run with the same 'tree' and different alignment options}
}
\value{
The same split table from getTreeSplits(tree), but with (1) no root row, and (2) a support column for the specfied alignment/missing combination
}
\description{
This function maps phylogenetic signal from a multiple-sequence alingment onto a rooted phylo or multiPhylo object
}
