% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTreeSupport.R
\name{getTreeSupport}
\alias{getTreeSupport}
\title{Get Alignment Support for Specific Topology}
\usage{
getTreeSupport(tree, signal, existing_splits)
}
\arguments{
\item{tree}{Rooted phylo object}

\item{signal}{Output table from getAlignmentSignal run with same taxa as in tree}

\item{existing_splits}{OPTIONAL: Should be output of getTreeSupport from the same tree, run with other alignments}
}
\value{
The same split table from getAllSplits(tree), but with a support column listing total site counts supporting each split
}
\description{
Given a (1) rooted tree, and
        (2) the signal output from getAlignmentSignal from an alignment and matched set of taxa, this function calculates alignment support for each split
}
\examples{
getTreeSupport(tree,signal)

}
