% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrimmedTree.R
\name{getTrimmedTree}
\alias{getTrimmedTree}
\title{Rboretum Tree Trimmer}
\usage{
getTrimmedTree(tree, taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{taxa}{Character vector of desired tip labels to keep}
}
\value{
Pruned phylo object
}
\description{
This ape wrapper returns a phylo object that has been pruned to include only specified taxa. Note: Function will STOP if tree is missing taxa.
}
\examples{
getTrimmedTree(tree,taxa)

}
