% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUniqueTopologies.R
\name{getUniqueTopologies}
\alias{getUniqueTopologies}
\title{Rboretum Unique Topology Fetcher}
\usage{
getUniqueTopologies(trees, return_table)
}
\arguments{
\item{trees}{Named, rooted multiPhylo object}

\item{return_table}{OPTIONAL: If TRUE, return summary table rather than multiPhylo}
}
\value{
multiPhylo containing unique topologies; or, summary table of the same data
}
\description{
Rboretum Unique Topology Fetcher
}
