% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCladeSupport.R
\name{isCladeSupport}
\alias{isCladeSupport}
\title{Rboretum Clade Support Checker}
\usage{
isCladeSupport(test_object, tree)
}
\arguments{
\item{test_object}{R object to check}

\item{tree}{Check that passed support object contains information about a specific set of trees. Options include:
\itemize{
  \item A rooted phylo object; or,
  \item A named, rooted multiPhylo object where all trees share 3+ taxa
}}
}
\value{
TRUE if output of getTreeClades(tree,return_count=TRUE); otherwise, FALSE
}
\description{
This function returns TRUE if all the clades in 'tree' are present in the clade comparsion output from getTreeClades(return_counts=TRUE); otherwise, FALSE
}
