% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMultiPhylo.R
\name{isMultiPhylo}
\alias{isMultiPhylo}
\title{Rboretum MultiPhylo Checker}
\usage{
isMultiPhylo(test_object, check_rooted)
}
\arguments{
\item{test_object}{R object to check}

\item{check_rooted}{OPTIONAL: If TRUE, also check if tree is rooted. Default: FALSE, don't check rootedness}
}
\value{
TRUE if multiPhylo with 2+ trees, otherwise FALSE
}
\description{
This function returns TRUE if the passed object is of class multiPhylo and has 2+ trees; Otherwise, FALSE
}
