% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMultiPhylo.R
\name{isMultiPhylo}
\alias{isMultiPhylo}
\title{Rboretum MultiPhylo Checker}
\usage{
isMultiPhylo(
  test_object,
  check_rooted,
  check_names,
  check_three_taxa,
  check_all_taxa,
  check_unique
)
}
\arguments{
\item{test_object}{R object to check}

\item{check_rooted}{OPTIONAL: If TRUE, also check if all trees are rooted. Default: FALSE, don't check rootedness}

\item{check_three_taxa}{OPTIONAL: If TRUE, also check if all trees share at least three taxa. Default: FALSE, don't check taxa}

\item{check_all_taxa}{OPTIONAL: If TRUE, also check if trees share identical taxa. Default: FALSE, don't check taxa}

\item{check_unique}{OPTIONAL: If TRUE, also check if trees all have a unique topology. Default: FALSE, don't check topology}

\item{check_named}{OPTIONAL: If TRUE, also check if trees are named. Default: FALSE, don't check names}
}
\value{
TRUE if all flagged conditions match, otherwise FALSE
}
\description{
This function returns TRUE if the passed object is of class multiPhylo and has 2+ trees; Otherwise, FALSE
}
