% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isMultiPhylo.R
\name{isMultiPhylo}
\alias{isMultiPhylo}
\title{Rboretum MultiPhylo Checker}
\usage{
isMultiPhylo(
  test_object,
  check_named,
  check_rooted,
  check_three_taxa,
  check_all_taxa,
  verbose
)
}
\arguments{
\item{test_object}{R object to check}

\item{check_named}{OPTIONAL: If TRUE, also check if trees are named. [Default: FALSE, don't check names]}

\item{check_rooted}{OPTIONAL: If TRUE, also check if all trees are rooted. [Default: FALSE, don't check rootedness]}

\item{check_three_taxa}{OPTIONAL: If TRUE, also check if all trees share at least three taxa. [Default: FALSE, don't check taxa]}

\item{check_all_taxa}{OPTIONAL: If TRUE, also check if all trees share all taxa. [Default: FALSE, don't check taxa]}

\item{verbose}{OPTIONAL: If TRUE: Print messages along the way; else execute silently. [Default: FALSE]}
}
\value{
TRUE if 'test_object' is a multiPhylo with 2+ trees; otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is of class multiPhylo and has 2+ trees; Otherwise, FALSE
}
