% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isPhylo.R
\name{isPhylo}
\alias{isPhylo}
\title{Rboretum Phylo Checker}
\usage{
isPhylo(test_object, check_rooted)
}
\arguments{
\item{test_object}{R object to check}

\item{check_rooted}{OPTIONAL: If TRUE, also check if tree is rooted. Default: FALSE, don't check rootedness}
}
\value{
TRUE if phylo, otherwise FALSE
}
\description{
This function returns TRUE if the passed object is of class phylo; Otherwise, FALSE
}
