% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTreeSupport.R
\name{isTreeSupport}
\alias{isTreeSupport}
\title{Rboretum Tree Support Checker}
\usage{
isTreeSupport(test_object, test_clade, partial)
}
\arguments{
\item{test_object}{R object to check}

\item{test_clade}{Source for clades to check. Function checks that passed support object contains information about these clades. Options include:
\itemize{
  \item Rooted phylo or multiPhylo object
  \item vector of semicolon-separated clades. 
}}

\item{partial}{OPTIONAL: If TRUE, as long as all clades from 'test_clade' are present, return TRUE [Default: FALSE, require that test_clades also includes all clades in support object]}
}
\value{
TRUE if output of getTreeSupport() with information on 'test_clade'; otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is the output of getTreeSupport() and contains information about clades in 'test_clade'; otherwise, FALSE
}
