% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTreeSupport.R
\name{isTreeSupport}
\alias{isTreeSupport}
\title{Rboretum Tree Support Checker}
\usage{
isTreeSupport(test_object, tree)
}
\arguments{
\item{test_object}{R object to check}

\item{tree}{OPTIONAL: phylo object [Check that passed tree support contains information about a specific tree.]}
}
\value{
TRUE if output of getTreeSupport(); otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is the output of getTreeSupport; otherwise, FALSE
}
