% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isTreeSupport.R
\name{isTreeSupport}
\alias{isTreeSupport}
\title{Rboretum Alignment Signal Checker}
\usage{
isTreeSupport(test_object, tree)
}
\arguments{
\item{test_object}{R object to check}

\item{tree}{OPTIONAL: Check that passed support object contains information about a specific set of trees. Options include:
\itemize{
  \item A rooted phylo object; or,
  \item A named, rooted multiPhylo object where all trees share 3+ taxa
}}
}
\value{
TRUE if output of getTreeSupport(); otherwise, FALSE
}
\description{
This function returns TRUE if the passed object is the output of getAlignmentSignal; otherwise, FALSE
}
