% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRooted_Worker.R
\name{readRooted_Worker}
\alias{readRooted_Worker}
\title{Rboretum Rooted Tree Reader}
\usage{
readRooted_Worker(to_root_worker, root_taxa, disable_bs)
}
\arguments{
\item{root_taxa}{Character vector containing outgroup species IDs (Must be in tree(s) and monophyletic)}

\item{disable_bs}{OPTIONAL: If TRUE, don't add a mirrored node label to unrooted trees [Default: FALSE; if trees are unrooted, mirror the missing node label]}

\item{to_root}{Where to find tree files. Options include:
\itemize{
  \item A character vector of one or more tree file paths
  \item A path to a single directory containing all tree files
}}
}
\value{
A phylo object, rooted at specified taxa
}
\description{
This function simultaenously reads in and roots one or more trees at a common root clade
}
