% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRooted_Worker.R
\name{readRooted_Worker}
\alias{readRooted_Worker}
\title{Rboretum Rooted Tree Reader Worker}
\usage{
readRooted_Worker(to_root_worker, root_taxa)
}
\arguments{
\item{to_root_worker}{Path to tree file}

\item{root_taxa}{Character vector containing outgroup species IDs (Must be in tree and monophyletic)}
}
\value{
A phylo object, rooted at specified taxa
}
\description{
This function reads in and roots a phylo object
}
