% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerootTree.R
\name{rerootTree}
\alias{rerootTree}
\title{Rboretum Tree Re-Rooter}
\usage{
newRootTaxa <- c('NewRoot_1','NewRoot_2')
rerootedTree  <- rerootTree(myTree,newRootTaxa)
}
\arguments{
\item{tree}{Phylo object}

\item{root_taxa}{Character vector containing outgroup species IDs (Must be in tree and monophyletic)}
}
\value{
A phylo object rooted at specified taxa
}
\description{
This ape wrapper accepts a phylo object and a character vector of new root taxa and returns a rooted phylo object, if possible
}
\examples{
rerootTree(birdTree,c('Alligator','Turtle'))
}
