% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescaleTreeSupport.R
\name{rescaleTreeSupport}
\alias{rescaleTreeSupport}
\title{Rboretum Tree Support Rescaler}
\usage{
rescaleTreeSupport(tree_support, scale)
}
\arguments{
\item{tree_support}{Output of getTreeSupport}

\item{scale}{Scaling factor. Options include:
\itemize{
  \item Single numeric value [return a repeated vector of this value (e.g. to set all geom_nodepoint to this size)]
  \item Two-item numeric vector [return summation values that have been rescaled to fit between c(min,max) (e.g. rescale millions of sites to range between c(1,20))]
  \item "log" [return log-transformed totals]
}}
}
\value{
Numeric vector the length of the number of clades, and corresponding to rescaled total support values
}
\description{
This function takes the output of getTreeSupport, and returns a numeric vector of total, summed support that has been rescaled
}
