% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.tree.R
\name{root.tree}
\alias{root.tree}
\title{Rboretum Tree Rooter}
\usage{
\method{root}{tree}(tree, root_taxa)
}
\arguments{
\item{tree}{Phylo object}

\item{root_taxa}{Character vector containing outgroup species IDs (Must be in tree and monophyletic)}
}
\value{
A phylo object rooted at specified taxa
}
\description{
This ape wrapper accepts a phylo object and a character vector of new root taxa and returns a rooted phylo object, if possible
}
\examples{
root.tree(birdTree,c('Alligator','Turtle'))
}
