% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameTopology.R
\name{sameTopology}
\alias{sameTopology}
\title{Rboretum Topology Checker}
\usage{
sameTopology(tree_1, tree_2)
}
\arguments{
\item{tree_1}{Phylo object}

\item{tree_2}{Phylo object}
}
\value{
TRUE if 'tree_1' and 'tree_2' have the same topology after pruning to identical species lists; else, FALSE
}
\description{
This function returns TRUE if 'tree_1' and 'tree_2' have the same topology after pruning to identical species lists
}
\examples{
sameTopology(tree_1,tree_2)

}
