% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sameTopology.R
\name{sameTopology}
\alias{sameTopology}
\title{Rboretum Identical Topology Checker}
\usage{
sameTopology(trees)
}
\arguments{
\item{trees}{multiPhylo object}
}
\value{
TRUE if all trees have the same topology after pruning to identical species lists; else, FALSE
}
\description{
This function returns TRUE if all trees in multiPhylo have the same topology after pruning to identical species lists
}
