% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiSorter.R
\name{semiSorter}
\alias{semiSorter}
\title{Rboretum Semicolon Sorter}
\usage{
semiSorter(string_to_sort)
}
\arguments{
\item{string_to_split}{Character vector}
}
\value{
Sorted semicolon delimited character vector
}
\description{
This function returns a sorted, semicolon-delimited character vector from a plain character vector, or a character vector of semicolon-delimited elements
}
\examples{
myShuffledString <- 'c;d;a;b'
semiSorter(myShuffledString)
> 'a;b;c;d'
}
