% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiVector.R
\name{semiVector}
\alias{semiVector}
\title{Create Vector from Semicolon Delimited String}
\usage{
semiVector(string_to_split)
}
\arguments{
\item{string_to_split}{Semicolon delimited string}
}
\value{
Vector of items
}
\description{
This function returns a vector of items from a string separated by semicolons (;)
}
\examples{
semiVector(string_to_split)

}
