% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiVector.R
\name{semiVector}
\alias{semiVector}
\title{Rboretum Semicolon Vectorizer}
\usage{
semiVector(string_to_split)
}
\arguments{
\item{string_to_split}{Semicolon delimited character vector of length 1}
}
\value{
Character vector
}
\description{
This function takes a string separated by semicolons (;) and returns a character vector
}
\examples{
myString <- 'a;b;c;d'
semiVector(myString)
> ['a','b','c','d']
}
