% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semiVector.R
\name{semiVector}
\alias{semiVector}
\title{Rboretum Semicolon Vectorizer}
\usage{
semiVector(string_to_split)
}
\arguments{
\item{string_to_split}{Semicolon delimited character}
}
\value{
If string_to_split is a single item, semiVector returns a character vector split by ';'. If string_to_split contains multiple items, function returns a list of separated vectors
}
\description{
This function takes a string separated by semicolons (;) and returns a character vector
}
\examples{
myString <- 'a;b;c;d'
semiVector(myString)
myStrings <- c('a;b;c;d','e;f;g;h')
semiVector(myStrings)
}
