% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceRboretum.R
\name{sourceRboretum}
\alias{sourceRboretum}
\title{Rboretum Python Script Sourcer}
\usage{
sourceRboretum()
}
\description{
This function sources all Rborteum Python scripts for 'reticulate', and globally assigns the test data directories
}
