% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCount.R
\name{tableCount}
\alias{tableCount}
\title{Search Table for Value}
\usage{
tableCount(table_to_search, name_to_search)
}
\arguments{
\item{table_to_search}{Table (Result of 'table' call)}

\item{name_to_search}{Title of item you want the count for}
}
\value{
Number of times that value occurs in the table
}
\description{
This function returns the count of items in a table, or 0 if item does not occur in table
}
\examples{
tableCount(table_to_search,name_to_search)

}
