% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandem.treePlot.R
\name{tandem.treePlot}
\alias{tandem.treePlot}
\title{Rboretum Simple Tandem Plotter
Simple cowplot wrapper to plot two ggtree/ggplot objects side by side}
\usage{
tandem.treePlot(plot1, plot2)
}
\arguments{
\item{plot1}{ggtree/ggplot object to be plotted on the left}

\item{plot2}{ggtree/ggplot object to be plotted on the right}
}
\value{
Side-by-side ggplot object
}
\description{
Rboretum Simple Tandem Plotter
Simple cowplot wrapper to plot two ggtree/ggplot objects side by side
}
\examples{
tandem.treePlot(ggtree1,ggtree2)
}
