% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tandemPlotter.R
\name{tandemPlotter}
\alias{tandemPlotter}
\title{Rboretum Tandem Plotter
Simple cowplot wrapper to plot two or more ggtree/ggplot objects side by side. Plots can be passed individually, or as a list}
\usage{
tandemPlotter(...)
}
\value{
Side-by-side ggplot object
}
\description{
Rboretum Tandem Plotter
Simple cowplot wrapper to plot two or more ggtree/ggplot objects side by side. Plots can be passed individually, or as a list
}
