% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_lm.R
\name{time_lm}
\alias{time_lm}
\title{Rboretum Time Linear Modeler}
\usage{
time_lm(age_support_df, lm_alpha)
}
\arguments{
\item{age_support_df}{}

\item{lm_alpha}{}
}
\value{

}
\description{
This function takes information about node ages and node support, and uses linear models and Bonferroni correction to assess changes in relative phylogenetic utility over time among datasets
}
