% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.support.R
\name{tree.support}
\alias{tree.support}
\title{Get Alignment Support for Specific Topology}
\usage{
tree.support(tree, signal, existing_splits, max_missing)
}
\arguments{
\item{tree}{Rooted phylo object}

\item{signal}{Output table from getAlignmentSignal run with same taxa as in tree}

\item{existing_splits}{OPTIONAL: Should be output of getTreeSupport from the same tree, run with other alignments}

\item{max_missing}{x}
}
\value{
The same split table from getAllSplits(tree), but with a support column listing total site counts supporting each split
}
\description{
Given a (1) rooted tree, and
        (2) the signal output from getAlignmentSignal from an alignment and matched set of taxa, this function calculates alignment support for each split
}
\examples{
tree.support(tree,signal)

}
