% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeDater.R
\name{treeDater}
\alias{treeDater}
\title{Rboretum Tree Dater}
\usage{
treeDater(tree, method, calibration_df, iterations)
}
\arguments{
\item{tree}{Tree(s) from which to extract branch length information. Options include:
\itemize{
  \item A single, rooted phylo object; or,
  \item A rooted multiPhylo object where all trees share all taxa and support a single topology 
}}

\item{method}{Estimation of node ages via:
\itemize{
  \item 'reltime_uncal'[Default]: A RelTime-like implementation without node age calibration (Relative dating)
  \item 'reltime': A RelTime-like implementation, allowing for relative-to-absolute age conversion via a single node calibration supplied by 'calibration_df' (if more than one calibration is provided, only the first entry will be processed.)
  \item 'chronos': Estimate node age using 'chronos' from 'ape', allowing for relative-to-absolute age conversion via multiple node calibrations supplied by 'calibration_df'
  }}

\item{calibration_df}{Required if method is 'reltime' or 'chronos'; Dataframe with 4 columns: (1) Taxon 1 (2) Taxon 2 (3) Min divergence time (4) Max divergence time}

\item{iterations}{If using 'chronos', how many times to estimate the age of each node prior to summarizing [Default: 1000]}
}
\value{
An ultrametric phylo object with branch lengths corresponding to time, or a multiPhylo of such trees
}
\description{
This function takes a tree(s) with estimated branch lengths, and infers either relative or absolute node ages
}
