% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treePlotter.R
\name{treePlotter}
\alias{treePlotter}
\title{Rboretum Tree Plotter}
\usage{
myPlot <- treePlotter(tree=myTree,...)
}
\arguments{
\item{tree}{Phylo object}

\item{branch_length}{OPTIONAL: TRUE [plot tree with branch lengths]; FALSE [Default: plot cladogram]}

\item{branch_weight}{OPTIONAL: Set ggtree branch thickness}

\item{node_label}{OPTIONAL: 'none' [no node labels]; 'node' [print node numbers]; 'bs' [Default: print node labels from tree (e.g. bootstrap)]}

\item{node_size}{OPTIONAL: Set ggtree node label size}

\item{node_nudge}{OPTIONAL: Set ggtree node label nudge_x}

\item{taxa_size}{OPTIONAL: Set ggtree tip label size}

\item{taxa_italic}{OPTIONAL: TRUE [italic tip labels]; FALSE [Default: non-italic tip labels]}

\item{taxa_align}{OPTIONAL: 'left' or 'right' tip label alignment [Default: No alignment]}

\item{taxa_offset}{OPTIONAL: Set ggtree tip label offset}

\item{xmax}{OPTIONAL: Set ggplot xlim upper limit (e.g if long tip labels run off plot)}
}
\value{
ggtree object
}
\description{
Given a phylo object (tree), this ggtree wrapper returns a ggtree plot object, adjusted with possible arguments
}
\examples{
# Print tree with bootstrap labels
treePlotter(tree)

# Print tree with no node labels
treePlotter(tree,node_label='none')

# Print tree with thicker branches
treePlotter(tree,branch_weight=3)

# Print tree with buffer on right side to accomodate longer tip labels
treePlotter(tree,xmax=10)
}
