% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treeTrimmer.R
\name{treeTrimmer}
\alias{treeTrimmer}
\title{Rboretum Tree Trimmer}
\usage{
treeTrimmer(tree, taxa, remove, keep_bs)
}
\arguments{
\item{tree}{Tree(s) to trim. Options include:
\itemize{
  \item A phylo object
  \item A multiPhylo object
}}

\item{taxa}{Character vector of desired tip labels to keep (or discard if remove=TRUE)}

\item{remove}{OPTIONAL: If TRUE, tip labels specified by 'taxa' are removed from all trees rather than retained [Default: FALSE, prune 'tree' down to 'taxa']}

\item{keep_bs}{OPTIONAL: If TRUE, do not remove bootstrap labels when taxa are trimmed off of tree [Default: FALSE, if tree has taxa removed, also remove node labels]}
}
\value{
Pruned phylo or multiPhylo object
}
\description{
This function returns a phylo or multiPhylo object that has been pruned to include only specified taxa. Note: Function will STOP if any 'tree' is missing any 'taxa' or if trimming will result in <2 final taxa
}
\examples{
# myTree is a phylo object with 4 tips
desiredTaxa <- c('Species_1','Species_2','Species_3')
undesiredTaxa <- 'Species_4'
trimmedTree <- treeTrimmer(myTree,desiredTaxa) # Trim myTree down to taxa in 'desiredTaxa'
trimmedTree <- treeTrimmer(myTree,undesiredTaxa,remove=TRUE) # Remove 'undesiredTaxa' from myTree
}
