% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorSemi.R
\name{vectorSemi}
\alias{vectorSemi}
\title{Rboretum Vector Semicolonizer}
\usage{
vectorSemi(to_semi)
}
\arguments{
\item{char_vec}{Character vector or list of character vectors}
}
\value{
Semicolon-separated strings for each passed vector
}
\description{
This function takes a character vector, or a list of vectors, and retuns a string separated by semicolons (;)
}
\examples{
myVec <- c('a','b','c','d')
vectorSemi(myVec)
> 'a;b;c;d'
}
